% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_site_EcoDataCube.R
\name{apply_color_table}
\alias{apply_color_table}
\title{Apply color table, and categories to raster from SLD file}
\usage{
apply_color_table(sld_url, r)
}
\arguments{
\item{sld_url}{Character full path to SLD file from EcoDataCube}

\item{r}{terra::rast The raster to apply color table}
}
\description{
Download SLD style file and apply color table to raster
}
\author{
Micha Silver, phD (2020) \email{silverm@post.bgu.ac.il}
}
