% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createQRCode.R
\name{createQRCode}
\alias{createQRCode}
\title{eLTER createQRCode function}
\usage{
createQRCode(deimsid)
}
\arguments{
\item{deimsid}{A `character`. It is the DEIMS iD of site make from
DEIMS-SDR website. More information about DEIMS iD in this
\href{https://deims.org/docs/deimsid.html}{page}.}
}
\value{
The output of the function is a QRcode `image`.
}
\description{
This function allows to obtain the QRCode of any DEIMS-SDR
entities of the eLTER.
}
\examples{
createQRCode(deimsid = "https://deims.org/f30007c4-8a6e-4f11-ab87-569db54638fe")

}
\author{
Alessandro Oggioni, phD (2020) \email{oggioni.a@irea.cnr.it}
}
