% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createSiteReport.R
\name{createSiteReport}
\alias{createSiteReport}
\title{eLTER_createSiteReport}
\usage{
createSiteReport(deimsid)
}
\arguments{
\item{deimsid}{is a DEIMS iD of site make from DEIMS-SDR website. More information about DEIMS iD in this page https://deims.org/docs/deimsid.html.}
}
\value{
The output of the function is a xxx with boundaries of the site.
}
\description{
This function allows to obtain a pdf report of eLTER site selected by DEIMS.ID. All the info are taken from the DEIMS-SDR sites API.
}
\examples{
createSiteReport(deimsid = 'https://deims.org/f30007c4-8a6e-4f11-ab87-569db54638fe')

}
\author{
Alessandro Oggioni, phD (2020) <oggioni.a@irea.cnr.it>
}
