% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_Q.R
\name{do_Q}
\alias{do_Q}
\title{eLTER do_Q function}
\usage{
do_Q(q, jj)
}
\arguments{
\item{q}{a character}

\item{jj}{a character}
}
\value{
The output of the function is ...
}
\description{
This function parses json content returned from an HTTP request
}
\examples{
deimsid <- "https://deims.org/f30007c4-8a6e-4f11-ab87-569db54638fe"
q <- '{title: .title,
  uri: "\\\\(.id.prefix)\\\\(.id.suffix)",
  geoCoord: .attributes.geographic.coordinates,
  country: .attributes.geographic.country,
  geoElev: .attributes.geographic.elevation,
  affiliation: .attributes.affiliation
}'
url <- paste0(
  "https://deims.org/",
  "api/sites/",
  sub("^.+/", "", deimsid)
)
export <- httr::GET(url = url)
jj <- suppressMessages(httr::content(export, "text"))

ReLTER:::do_Q(q, jj)

}
\author{
Paolo Tagliolato, phD (2021) \email{tagliolato.p@irea.cnr.it}
}
\keyword{internal}
