% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eLTER_reporting.R
\name{elter_write_rdata}
\alias{elter_write_rdata}
\title{eLTER write rds data function}
\usage{
elter_write_rdata(myfiles, delim)
}
\arguments{
\item{myfiles}{A `character`. The list of the files to deposit in Zenodo.
Please provide all files only with 'csv' extension.}

\item{delim}{A `character`. Provide the character used to separate fields
within a record. Only if the extension of the file(s) are 'csv', 'tsv', or
'txt'.}
}
\value{
This function returns a rds files.
}
\description{
This function write a rds file from csv, tsv, txt, xls or xlsx
dataset
}
\examples{
elter_write_rdata(
  myfiles = c(
   "miscellaneus/file_show/data_mapping.csv",
   "miscellaneus/file_show/reference_TAXA.csv",
   "miscellaneus/file_show/reference_VARIABLES.csv"
  ),
  delim = ";"
)

}
\author{
Alessandro Oggioni, phD \email{oggioni.a@irea.cnr.it}
}
\keyword{internal}
