% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produceMapOfSiteFromDEIMS.R
\name{fProduceMapOfSiteFromDEIMS}
\alias{fProduceMapOfSiteFromDEIMS}
\title{eLTER_spatial_points_of_sites}
\usage{
fProduceMapOfSiteFromDEIMS(
  deimsid,
  countryCode,
  listOfSites,
  gridNx,
  gridNy,
  width,
  height,
  siteName,
  bboxXMin,
  bboxXMax,
  bboxYMin,
  bboxYMax
)
}
\arguments{
\item{networkDEIMSID}{is a DEIMS iD of network make from DEIMS-SDR website. More information about DEIMS iD in this page https://deims.org/docs/deimsid.html, and at this page https://deims.org/search?f%5B0%5D=result_type%3Anetwork the complete list of iLTER networks.}
}
\value{
The output of the function is SpatialPointsDataFrame (package sp).
}
\description{
This function provide a SpatialPointsDataFrame (sp) of sites in a network provided by DEIMS-SDR.
}
\examples{
sitesNetwork <- ReLTER::getNetworkSites(networkDEIMSID = 'https://deims.org/network/7fef6b73-e5cb-4cd2-b438-ed32eb1504b3')
sitesNetwork <- (sitesNetwork[!grepl('^IT', sitesNetwork$title),])
fProduceMapOfSiteFromDEIMS(
  deimsid = 'https://deims.org/f30007c4-8a6e-4f11-ab87-569db54638fe',
  countryCode = 'ITA',
  listOfSites = sitesNetwork,
  gridNx = 0.7,
  gridNy = 0.35,
  width = 0.25,
  height = 0.25,
  siteName = 'Lago Maggiore',
  bboxXMax = 0,
  bboxXMin = 0,
  bboxYMin = 0,
  bboxYMax = 0
)

}
\author{
Alessandro Oggioni, phD (2020) <oggioni.a@irea.cnr.it>
}
