% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset_info.R
\name{get_dataset_info}
\alias{get_dataset_info}
\title{Return a tibble object containing information, as a stored in
\href{https://deims.org/}{DEIMS-SDR catalogue}, about a single eLTER dataset
(e.g. \url{https://deims.org/dataset/38d604ef-decb-4d67-8ac3-cc843d10d3ef}).}
\usage{
get_dataset_info(datasetid)
}
\arguments{
\item{datasetid}{A character. It is the DEIMS.iD of dataset make from
DEIMS-SDR website. More information about DEIMS.iD in this
\href{https://deims.org/docs/deimsid.html}{page}.}
}
\value{
The output of the function is a `tibble` with main features
of the site and the related resources collected by site.
}
\description{
This function allows to obtain the info of dataset provided in
DEIMS-SDR.
}
\examples{
tDataset <- get_dataset_info(
  datasetid =
  "https://deims.org/dataset/38d604ef-decb-4d67-8ac3-cc843d10d3ef"
)
tDataset

}
\author{
Alessandro Oggioni, phD (2020) \email{oggioni.a@irea.cnr.it}
}
