% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_deims_api_version.R
\name{get_deims_API_version}
\alias{get_deims_API_version}
\title{Get version of DEIMS-SDR API}
\usage{
get_deims_API_version(deims_url = get_deims_base_url())
}
\arguments{
\item{deims_url}{A \code{character}. DEIMS-SDR base URL. Defaults to package
settings.}
}
\value{
version number.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
This function obtains the version of the DEIMS-SDR API.
}
\references{
\insertRef{httr2R}{ReLTER}

\insertRef{dplyrR}{ReLTER}
}
