% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{get_deims_base_url}
\alias{get_deims_base_url}
\title{Get DEIMS-SDR base URL}
\usage{
get_deims_base_url()
}
\value{
DEIMS-SDR base URL
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\seealso{
Other package_customizable_settings:
\code{\link{package_settings}}
}
\concept{package_customizable_settings}
