% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_id.R
\name{get_id}
\alias{get_id}
\title{eLTER get_id function}
\usage{
get_id(deimsid, resource = "sites", test = NULL, ...)
}
\arguments{
\item{deimsid}{A \code{character}. It is the DEIMS ID of the site, activity or
dataset from DEIMS-SDR website. DEIMS ID information
\href{https://deims.org/docs/deimsid.html}{here}.}

\item{resource}{Character: one among \code{"sites"} (default), \code{"activities"} or
\code{"datasets"} (\code{"networks"} currently not tested).}

\item{test}{Logical: if TRUE, content is not retrieved from
\url{https://deims.org} but internally (only for some defined IDs).
Used for testing purposes to avoid errors in case or HTTP error 500.
If missing, the content of the internal variable}

\item{...}{Arguments to be passed to \code{httr::RETRY()}.}
}
\value{
A character containing the json content, and an attribute \code{status}.
}
\description{
Internal function to retrieve json content from ID
}
\author{
Alessandro Oggioni, phD (2020) \email{oggioni.a@irea.cnr.it}

Luigi Ranghetti, phD (2021) \email{luigi@ranghetti.info}

Paolo Tagliolato, PhD (2021) \email{tagliolato.p@irea.cnr.it}
}
\keyword{internal}
