% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ilter_observedProperties.R
\name{get_ilter_observedProperties}
\alias{get_ilter_observedProperties}
\title{Obtain information about the observed properties collected of all sites
ILTER.}
\usage{
get_ilter_observedProperties(sitesNum = 0)
}
\arguments{
\item{sitesNum}{A \code{integer}. The number of the sites that are
read to get the information. Use this parameter only to sample the output
of this function. If the value of sitesNum is #' 0 (default)
all the ILTER sites will be parsed and the waiting time will be long.}
}
\value{
The output of the function is a \code{tibble} containing the list
of observed properties and their URI (Uniform Resource Identifier) collected
in all ILTER sites.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#questioning}{\figure{lifecycle-questioning.svg}{options: alt='[Questioning]'}}}{\strong{[Questioning]}}
Return a \code{tibble} object containing observed properties
collected by all
of the \href{https://www.ilter.network/network/global-coverage}{ILTER sites
(more than 1200 around the world)}, available from
\href{https://deims.org}{DEIMS-SDR}.

This function gathers in a unique tibble all the observed properties
from all ILTER sites. Note that the execution time for this function
is very high.

If the objective is obtain information about observed properties
on a few sites, it is better to use other more specific functions (e.g.
\href{get_network_observedProperties.html}{\code{get_network_observedProperties()}}
or \href{get_site_info.html}{\code{get_site_info()}}) or using
other methods
(\href{../../articles/sites_information.html}{How to about sites
informations}).
}
\examples{
\dontrun{
listParams <- get_ilter_observedProperties(sitesNum = 20)
listParams[1:10, ] \%>\%
  dplyr::rows_insert(
  dplyr::tibble(
    parameterLabel = "...", parameterUri = "..."
  )
)
}

}
\author{
Alessandro Oggioni, phD (2020) \email{oggioni.a@irea.cnr.it}
}
