% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ilter_parameters.R
\name{get_ilter_parameters}
\alias{get_ilter_parameters}
\title{eLTER get_ilter_parameters function}
\usage{
get_ilter_parameters(sitesNum = 0)
}
\arguments{
\item{sitesNum}{A `integer`. It is the number of the sites that are
read to get the information. Use this parameters moreover for provide
example of this function. Default 0.}
}
\value{
The output of the function is a `tibble` containing the list
of parameters and their URI (Uniform Resource Identifier) collected
in all ILTER sites.
}
\description{
This function obtains information about the
parameters collected in ILTER sites, through the DEIMS-SDR sites API.
}
\examples{
\dontrun{
listParams <- get_ilter_parameters(sitesNum = 20)
listParams[1:10, ] \%>\%
  dplyr::rows_insert(
  dplyr::tibble(
    parameterLabel = "...", parameterUri = "..."
  )
)
}

}
\author{
Alessandro Oggioni, phD (2020) \email{oggioni.a@irea.cnr.it}
}
