% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_network_envcharacts.R
\name{get_network_envcharacts}
\alias{get_network_envcharacts}
\title{Obtain a list of all Environmental Characteristics of sites in an
eLTER Network.}
\usage{
get_network_envcharacts(networkDEIMSID)
}
\arguments{
\item{networkDEIMSID}{A `character`. DEIMS.iD of network
from DEIMS-SDR website. DEIMS.iD information 
\href{https://deims.org/docs/deimsid.html}{here} and Complete list of ILTER
networks \href{https://deims.org/search?f[0]=result_type:network}{here}.
The DEIMS.iD of network is the URL for the network page.}
}
\value{
The output of the function is a `tibble` containing all the
Environmental Characteristics of the network's sites.
}
\description{
This function obtains all Environmental Characteristics:
title, URI, geo-coordinates, country name, and elevation
of eLTER Network sites (e.g.
\href{https://deims.org/networks/7fef6b73-e5cb-4cd2-b438-ed32eb1504b3}{LTER-
Italy network}), through the DEIMS-SDR API.
}
\examples{
\dontrun{
listEnvCharacts <- get_network_envcharacts(
  networkDEIMSID =
  "https://deims.org/network/7fef6b73-e5cb-4cd2-b438-ed32eb1504b3"
)
listEnvCharacts[1:10, ]
}

}
\author{
Alessandro Oggioni, phD (2020) \email{oggioni.a@irea.cnr.it}
}
