% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_network_observedProperties.R
\name{get_network_observedProperties}
\alias{get_network_observedProperties}
\title{Obtain a list of all the observed properties of sites in an
eLTER Network.}
\usage{
get_network_observedProperties(networkDEIMSID)
}
\arguments{
\item{networkDEIMSID}{A \code{character}. The DEIMS ID of network
from DEIMS-SDR website. DEIMS ID information
\href{https://deims.org/docs/deimsid.html}{here} and Complete list of ILTER
networks \href{https://deims.org/search?f\link{0}=result_type:network}{here}.
The DEIMS ID of network is the URL for the network page.}
}
\value{
The output of the function is a \code{tibble} containing the list
of observed properties and their URI (Uniform Resource Identifier) collected
by the network's sites.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
This function obtains all observed properties
collected in an eLTER Network (e.g.
\href{https://deims.org/networks/7fef6b73-e5cb-4cd2-b438-ed32eb1504b3}{LTER-
Italy network}), through the DEIMS-SDR API.
}
\examples{
\dontrun{
listParams <- get_network_observedProperties(
  networkDEIMSID =
  "https://deims.org/networks/e0f680c2-22b1-4424-bf54-58aa9b7476a0"
)
listParams[1:10, ] \%>\%
dplyr::rows_insert(
  dplyr::tibble(parameterLabel = "...", parameterUri = "...")
)
}

}
\references{
\insertRef{jsonliteR}{ReLTER}

\insertRef{dplyrR}{ReLTER}
}
\author{
Alessandro Oggioni, phD (2020) \email{oggioni.a@irea.cnr.it}
}
