% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_network_parameters.R
\name{get_network_parameters}
\alias{get_network_parameters}
\title{eLTER get_network_parameters function}
\usage{
get_network_parameters(networkDEIMSID)
}
\arguments{
\item{networkDEIMSID}{A `character`. The DEIMS ID of network
from DEIMS-SDR website. For more information about DEIMS ID refer to these
pages: \href{https://deims.org/docs/deimsid.html}{page}, and
\href{https://deims.org/search?f[0]=result_type:network}{page}
the complete list of ILTER networks.}
}
\value{
The output of the function is a `tibble` containing the list
of parameters and their URI (Uniform Resource Identifier) collected
by the network's sites.
}
\description{
This function obtains a list of the environmental parameters
collected on sites in the eLTER Network (e.g. Italy) through the DEIMS-SDR
sites API.
}
\examples{
\dontrun{
listParams <- get_network_parameters(
  networkDEIMSID =
  "https://deims.org/networks/e0f680c2-22b1-4424-bf54-58aa9b7476a0"
)
listParams[1:10, ] \%>\%
dplyr::rows_insert(
  dplyr::tibble(parameterLabel = "...", parameterUri = "...")
)
}

}
\author{
Alessandro Oggioni, phD (2020) \email{oggioni.a@irea.cnr.it}
}
