% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_network_related_resources.R
\name{get_network_related_resources}
\alias{get_network_related_resources}
\title{Return a tibble object containing Related Resources, as a stored in
\href{https://deims.org/}{DEIMS-SDR catalogue}, of all the eLTER sites
belonging to an eLTER Network (e.g.
\href{https://deims.org/networks/7fef6b73-e5cb-4cd2-b438-ed32eb1504b3}{LTER
Italy network}).}
\usage{
get_network_related_resources(networkDEIMSID)
}
\arguments{
\item{networkDEIMSID}{A `character`. It is the DEIMS.iD of network make
from DEIMS-SDR website. DEIMS.iD information 
\href{https://deims.org/docs/deimsid.html}{here} and Complete list of ILTER
networks \href{https://deims.org/search?f[0]=result_type:network}{here}.
The DEIMS.iD of network is the URL for the network page.}
}
\value{
The output of the function is a `tibble` containing the related
resources shared by the network's sites.
}
\description{
This function obtains information about the
Related Resources (e.g. dataset) shared in the eLTER Network (e.g. Italy)
through the DEIMS-SDR sites API.
}
\examples{
\dontrun{
listRelatedResources <- get_network_related_resources(
  networkDEIMSID =
  "https://deims.org/network/7fef6b73-e5cb-4cd2-b438-ed32eb1504b3"
)
listRelatedResources[1:10, ]
}

}
\author{
Alessandro Oggioni, phD (2020) \email{oggioni.a@irea.cnr.it}
}
