% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_network_research_topics.R
\name{get_network_research_topics}
\alias{get_network_research_topics}
\title{Return a tibble object containing Research Topics collected by all
of the eLTER sites belonging to an eLTER Network (e.g.
\href{https://deims.org/networks/7fef6b73-e5cb-4cd2-b438-ed32eb1504b3}{LTER
Italy network}), as a stored into \href{https://deims.org}{DEIMS-SDR}.}
\usage{
get_network_research_topics(networkDEIMSID)
}
\arguments{
\item{networkDEIMSID}{A `character`. The DEIMS.iD of a network
from DEIMS-SDR website. More information about DEIMS network ID in these
pages: \href{https://deims.org/docs/deimsid.html}{page}, and
\href{https://deims.org/search?f[0]=result_type:network}{page} the
complete list of ILTER networks.}
}
\value{
The output of the function is a `tibble` containing the research
topics and their URI (Uniform Resource Identifier) collected by network's
sites.
}
\description{
This function obtains a list of Research Topics
handled in the eLTER Network (e.g. Italy) through
the DEIMS-SDR sites API.
}
\examples{
\dontrun{
listResearchTopics <- get_network_research_topics(
  networkDEIMSID =
  "https://deims.org/network/7fef6b73-e5cb-4cd2-b438-ed32eb1504b3"
)
listResearchTopics[1:10, ] \%>\%
  dplyr::rows_insert(
    dplyr::tibble(
      researchTopicsLabel = "...",
      researchTopicsUri = "..."
    )
  )
}

}
\author{
Alessandro Oggioni, phD (2020) \email{oggioni.a@irea.cnr.it}
}
