% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sos_observed_properties.R
\name{get_sensor_observed_properties}
\alias{get_sensor_observed_properties}
\title{Obtain the observed properties measured by a sensor.}
\usage{
get_sensor_observed_properties(sosURL, procedure)
}
\arguments{
\item{sosURL}{A `character`. The endpoint of the Sensor Observation Service
(SOS) service.}

\item{procedure}{A `character`. It is a procedure/sensor ID.}
}
\value{
The output of the function is a `tibble` with the labels and URI
(Uniform Resource Identifier) of each observed property, the code and
URI of Units Of Measurement (UOM) of the observed properties as declared
in the Sensor Observation Service (SOS). Codes and URIs as stated in
\href{https://qudt.org}{QUDT.org} are also present.
QUDT is a public charity nonprofit organization founded to provide
semantic specifications for units of measure, quantity kind, dimensions
and data types.
NB this function returns a valued string only in the case where the
UOM refers to a NERC vocabulary term (e.g.
http://vocab.nerc.ac.uk/collection/P06/current/UPAA/ for °C).
}
\description{
`r lifecycle::badge("experimental")`
This function obtains the observed properties by
procedure/sensor through Sensor Observation Service (SOS).
}
\examples{
FP <- get_sensor_observed_properties(
  sosURL = "http://getit.lteritalia.it/observations/service",
  procedure = "http://www.get-it.it/sensors/getit.lteritalia.it/procedure/noOwnerDeclared/noModelDeclared/noSerialNumberDeclared/1286194C-A5DF-11DF-8ED7-1602DFD72097"
)
FP

eurac_monalisa <- get_sensor_observed_properties(
  sosURL = "http://monalisasos.eurac.edu/sos/service",
  procedure = "QuantumSensor_nemef2000"
)
eurac_monalisa

# umweltbundesamt <- get_sensor_observed_properties(
#   sosURL = "https://ltercwn.umweltbundesamt.at/cwn-sos2/service",
#   procedure = "KAS_M"
# )
# The value 'D:/deployment/projects/cwnsos2node0140100/Tomcat/sensorML/sos2/sml2/KAS_M.xml' of the parameter 'procedure' is invalid


obsProsAir <- get_sensor_observed_properties(
  sosURL = "http://getit.lteritalia.it/observations/service",
  procedure = "http://www.get-it.it/sensors/getit.lteritalia.it/procedure/noOwnerDeclared/noModelDeclared/noSerialNumberDeclared/SI000049-1675AirTemp"
)
obsProsAir

NIVA <- get_sensor_observed_properties(
  sosURL = "https://hydro-sos.niwa.co.nz/",
  procedure = "Hydrometric_Station"
)
NIVA

}
\author{
Alessandro Oggioni, phD \email{oggioni.a@irea.cnr.it}
}
