% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_site_affiliations.R
\name{get_site_affiliations}
\alias{get_site_affiliations}
\title{eLTER get_site_affiliations function}
\usage{
get_site_affiliations(deimsid)
}
\arguments{
\item{deimsid}{A character. The DEIMS ID of the site from
DEIMS-SDR website. DEIMS ID information
\href{https://deims.org/docs/deimsid.html}{here}.}
}
\value{
The output of the function is a \code{tibble} with main features of the
site and the affiliations information, such as: networks and projects in
which the site is involved.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
This function obtains details about an eLTER site
through the DEIMS-SDR sites API.
}
\author{
Alessandro Oggioni, phD (2020) \email{oggioni.a@irea.cnr.it}
}
\keyword{internal}
