% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_site_infrastructure.R
\name{get_site_infrastructure}
\alias{get_site_infrastructure}
\title{eLTER get_site_infrastructure function}
\usage{
get_site_infrastructure(deimsid)
}
\arguments{
\item{deimsid}{A `character`. The DEIMS ID of the site from
DEIMS-SDR website. DEIMS ID information
\href{https://deims.org/docs/deimsid.html}{here}.}
}
\value{
The output of the function is a `tibble` with main features of the
site and infrastructure information where available, such as:
power supply, accessibility, maintenaince interval, etc.
}
\description{
This function obtains infrastructure information
of an eLTER site through the DEIMS-SDR sites API.
}
\author{
Alessandro Oggioni, phD (2020) \email{oggioni.a@irea.cnr.it}
}
\keyword{internal}
