% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_site_MODIS.R
\name{get_site_MODIS}
\alias{get_site_MODIS}
\title{Acquire a time series either of two datasets from MODIS satellites:
Land Surface Temperature (LST) or Vegetation Index (NDVI)}
\usage{
get_site_MODIS(
  deimsid,
  dataset = "VI",
  from_date = "2010.01.01",
  to_date = "2020.31.12",
  output_dir = NULL,
  plot_ts = TRUE
)
}
\arguments{
\item{deimsid}{A \code{character}. The DEIMS ID of the site from
DEIMS-SDR website. DEIMS ID information
\href{https://deims.org/docs/deimsid.html}{here}.}

\item{dataset}{A \code{character}. The requested dataset. One of:
"LST_day", "LST_night", VI"
Default is "VI".}

\item{from_date}{a \code{character}, the start date formatted as YYYY.MM.DD}

\item{output_dir}{a \code{character}, where to save downloaded rasters}

\item{plot_ts}{a \code{boolean}, whether to plot the time series,
default TRUE}

\item{to_data}{a \code{character}, the end date formatted as YYYY.MM.DD}
}
\value{
Full path to the downloaded Geotiff rasters.
This folder will contain requested datasets, cropped to the site boundaries.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Download a timeseries of MODIS images containing the requested
dataset and optionally plot a time series graph
of the average values over the site.
Use of this function requires registering on the EarthData website:
Requires registration on EarthData website:
https://urs.earthdata.nasa.gov/home
In order to guard your user credentials, please save
your username and password to environment variables. i.e.
Sys.setenv(earthdata_user="homer_simpson")
Sys.setenv(earthdata_pass="bart&lucy")
}
\details{
cropped to an eLTER site boundary.

Certain layers from one of these MODIS products are acquired.
from:
"LST_3band_emissivity_Daily_1km (M\emph{D21A1D)" and
"LST_3band_emissivity_Daily_1km_night (M}D21A1N)"
the "Land surface temperature" band is acquired.
from:
"Vegetation Indexes_16Days_250m (M*D13Q1)"
two bands are acquired: "16 day NDVI average" and "16 day EVI average"
NOTE that the default \code{output_dir} is tempdir(), so the downloaded files
will be deleted when exiting R.
}
\examples{
 \dontrun{

}

}
\references{
\insertRef{sfR}{ReLTER}

\insertRef{terraR}{ReLTER}
}
\author{
Micha Silver, phD (2020) \email{silverm@post.bgu.ac.il}

Alessandro Oggioni, phD (2020) \email{oggioni.a@irea.cnr.it}
}
