% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_site_parameters.R
\name{get_site_parameters}
\alias{get_site_parameters}
\title{eLTER get_site_parameters function}
\usage{
get_site_parameters(deimsid)
}
\arguments{
\item{deimsid}{A `character`. It is the DEIMS ID of the site from
DEIMS-SDR website. More information about DEIMS ID from:
\href{https://deims.org/docs/deimsid.html}{page}.}
}
\value{
The output of the function is a `tibble` with main features of the
site and the parameters collected.
}
\description{
This function obtains the list of parameters measured
in the eLTER site through the DEIMS-SDR sites API.
}
\author{
Alessandro Oggioni, phD (2020) \email{oggioni.a@irea.cnr.it}
}
\keyword{internal}
