% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_site_related_resources.R
\name{get_site_related_resources}
\alias{get_site_related_resources}
\title{eLTER get_site_related_resources function}
\usage{
get_site_related_resources(deimsid)
}
\arguments{
\item{deimsid}{A `character`. The DEIMS.iD of the site from
DEIMS-SDR website. DEIMS.iD information 
\href{https://deims.org/docs/deimsid.html}{here}.}
}
\value{
The output of the function is a `tibble` with main features of the
site and a list of the related resources collected by site.
}
\description{
This function obtains a list of related resources
collected in an eLTER site through the DEIMS-SDR sites API.
}
\author{
Alessandro Oggioni, phD (2020) \email{oggioni.a@irea.cnr.it}
}
\keyword{internal}
