% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_site_research_topics.R
\name{get_site_research_topics}
\alias{get_site_research_topics}
\title{eLTER get_site_research_topics function}
\usage{
get_site_research_topics(deimsid)
}
\arguments{
\item{deimsid}{A `character`. The DEIMS ID of the site from
DEIMS-SDR website. DEIMS ID information
\href{https://deims.org/docs/deimsid.html}{here}.}
}
\value{
The output of the function is a `tibble` with main features of the
site and a list of the research topics handled in this site.
}
\description{
This function obtains a list of research topics handled
at an eLTER site through the DEIMS-SDR sites API.
}
\author{
Alessandro Oggioni, phD (2020) \email{oggioni.a@irea.cnr.it}
}
\keyword{internal}
