% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sos_foi.R
\name{get_sos_foi}
\alias{get_sos_foi}
\title{Obtain the information about the feature of interest.}
\usage{
get_sos_foi(sosURL, show_map = FALSE)
}
\arguments{
\item{sosURL}{A `character`. The endpoint of the Sensor Observation Service
(SOS) service.}

\item{show_map}{A `logical`. When TRUE the boundary will be plotted on a
Leaflet map. Default FALSE.}
}
\value{
The output of the function is a `sf`. The table contains all the information
about Feature(s) Of Interest (FOI(s)) provided by a Sensor Observation
Services (SOS).
The columns are about: sampling feature (typeSf,
description, name), sampled feature id (sampledFeature), coordinate reference
system code (srsName) and coordinates (geometry).
}
\description{
`r lifecycle::badge("experimental")`
This function obtains the information about Feature(s) Of Interest (FOI(s))
provided by a Sensor Observation Services (SOS).
}
\examples{
LTERItaly <- get_sos_foi(
  sosURL = "http://getit.lteritalia.it/observations/service",
  show_map = TRUE
)
LTERItaly

eurac_monalisa <- get_sos_foi(
  sosURL = "http://monalisasos.eurac.edu/sos/service",
  show_map = TRUE
)
eurac_monalisa

# umweltbundesamt <- get_sos_foi(
#   sosURL = "https://ltercwn.umweltbundesamt.at/cwn-sos2/service"
# )
# The value 'D:/deployment/projects/cwnsos2node0140100/Tomcat/sensorML/sos2/sml2/KAS_M.xml' of the parameter 'procedure' is invalid


NIVA <- get_sos_foi(
  sosURL = "https://hydro-sos.niwa.co.nz/",
  show_map = TRUE
)
NIVA

}
\author{
Alessandro Oggioni, phD \email{oggioni.a@irea.cnr.it}

Paolo Tagliolato, phD \email{tagliolato.p@irea.cnr.it}
}
