% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sos_procedure_info.R
\name{get_sos_procedure_info}
\alias{get_sos_procedure_info}
\title{Obtain the information from a sensor.}
\usage{
get_sos_procedure_info(sosURL, procedure)
}
\arguments{
\item{sosURL}{A \code{character}. The endpoint of the Sensor Observation Service
(SOS) service.}

\item{procedure}{A \code{character}. It is a procedure/sensor ID.}
}
\value{
The output of the function is a \code{character} containing attributes
such as name of the sensor, id and their description.
All the information are collected by requests to Sensor Observation
Service (SOS).
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function obtains the information (metadata) shared by
procedure/sensor through Sensor Observation Service (SOS).
}
\examples{
FP <- get_sos_procedure_info(
  sosURL = "http://getit.lteritalia.it/observations/service",
  procedure = "http://www.get-it.it/sensors/getit.lteritalia.it/procedure/noOwnerDeclared/noModelDeclared/noSerialNumberDeclared/1286194C-A5DF-11DF-8ED7-1602DFD72097"
)
FP


# EURAC (https://www.eurac.edu/) 
# eurac_monalisa <- get_sos_procedure_info(
#   sosURL = "http://monalisasos.eurac.edu/sos/service",
#   procedure = "QuantumSensor_nemef2000"
# )
# eurac_monalisa

# obsProsAir <- get_sos_procedure_info(
#   sosURL = "http://getit.lteritalia.it/observations/service",
#   procedure = "http://www.get-it.it/sensors/getit.lteritalia.it/procedure/noOwnerDeclared/noModelDeclared/noSerialNumberDeclared/SI000049-1675AirTemp"
# )
# obsProsAir


# NIVA (https://niwa.co.nz)
# NIVA <- get_sos_procedure_info(
#   sosURL = "https://hydro-sos.niwa.co.nz/",
#   procedure = "Water_Quality_Site"
# )
# NIVA

}
\author{
Alessandro Oggioni, phD\email{oggioni.a@irea.cnr.it}

Paolo Tagliolato, phD \email{tagliolato.p@irea.cnr.it}
}
