% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sos_procedurelist.R
\name{get_sos_procedurelist}
\alias{get_sos_procedurelist}
\title{eLTER get_sos_procedurelist function}
\usage{
get_sos_procedurelist(sosURL)
}
\arguments{
\item{sosURL}{A `character`. The endpoint of the Sensor Observation Service
(SOS) service.}
}
\value{
The output of the function is a `list` with the name and URI (Uniform
Resource Identifier) of each procedure.
}
\description{
This function lists the procedures
of a Sensor Observations Service (SOS).
}
\examples{
\dontrun{
get_sos_procedurelist(
  sosURL = "http://getit.lteritalia.it/observations/service"
)
}

}
\author{
Alessandro Oggioni, phD (2020) \email{oggioni.a@irea.cnr.it}
}
