% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{getDeimsBaseUrl}
\alias{getDeimsBaseUrl}
\title{get deims base url}
\usage{
getDeimsBaseUrl()
}
\value{
deims base url
}
\description{
get deims base url
}
\seealso{
Other package_customizable_settings: 
\code{\link{package_settings}}
}
\concept{package_customizable_settings}
