% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDeimsSiteID.R
\name{getDeimsSiteID}
\alias{getDeimsSiteID}
\title{eLTER getDeimsSiteID function}
\usage{
getDeimsSiteID(sites_filtered, idx = 1)
}
\arguments{
\item{sites_filtered}{A `data.frame`.}

\item{idx}{An `integer` index in the data.frame for the desired site id}
}
\value{
the deims site id (url)
}
\description{
This function retrieves the DEIMS site ID
from a filtered list (data.frame) of sites (i.e. using getSitesByCountry)
where the user supplies an index number to choose which site from the list
}
\examples{
deimsid <- getDeimsSitesID(sites_filtered, 3)
deimsid
}
\author{
Micha Silver, \email{silverm@post.bgu.ac.il}
}
