% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getILTERGeneralInfo.R
\name{getILTERGeneralInfo}
\alias{getILTERGeneralInfo}
\title{eLTER getILTERGeneralInfo function}
\usage{
getILTERGeneralInfo(sitesNum = 0)
}
\arguments{
\item{sitesNum}{A `integer`. It is the number of the sites that are 
read to get the information. Use this parameters moreover for provide
example of this function. Default 0.}
}
\value{
The output of the function is a `sf` containing the
name, longitude, latitude, average altitude, DEIMS.iD and affiliation
of all ILTER sites.
}
\description{
This function allows to obtain the generic information
collected in all sites of ILTER, through the DEIMS-SDR
sites API.
}
\examples{
\dontrun{
require('dplyr')
listOfAllSites <- getILTERGeneralInfo(sitesNum = 20)
listOfAllSites[1:10, ]
}

}
\author{
Alessandro Oggioni, phD (2021) \email{oggioni.a@irea.cnr.it}
}
