% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getILTERParameters.R
\name{getILTERParameters}
\alias{getILTERParameters}
\title{eLTER getILTERParameters function}
\usage{
getILTERParameters(sitesNum = NULL)
}
\arguments{
\item{sitesNum}{A `integer`. It is the number of the sites that are 
read to get the information. Use this parameters oreover for provide
example of this function.}
}
\value{
The output of the function is a `tibble` containing the list
of parameters and their URI (Uniform Resource Identifier) collected
in all ILTER sites.
}
\description{
This function allows to obtain the information about the
parameters collected in all sites of ILTER, througth the DEIMS-SDR
sites API.
}
\examples{
\donttest
require('dplyr')
listParams <- getILTERParameters(sitesNum = 20)
listParams[1:10, ] \%>\%
  dplyr::rows_insert(
  dplyr::tibble(
    parameterLabel = "...", parameterUri = "..."
  )
)
\donttest

}
\author{
Alessandro Oggioni, phD (2020) \email{oggioni.a@irea.cnr.it}
}
