% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getILTERResearchTopics.R
\name{getILTERResearchTopics}
\alias{getILTERResearchTopics}
\title{eLTER_getILTERResearchTopics}
\usage{
getILTERResearchTopics()
}
\value{
The output of the function is ...
}
\description{
This function ...
}
\examples{
listResearchTopics <- getILTERResearchTopics()
knitr::kable(listResearchTopics[1:10, ] \%>\% dplyr::rows_insert(tibble(researchTopicsLabel = "...", researchTopicsUri = "...")))

}
\author{
Alessandro Oggioni, phD (2020) <oggioni.a@irea.cnr.it>
}
