% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNetworkEnvCharacts.R
\name{getNetworkEnvCharacts}
\alias{getNetworkEnvCharacts}
\title{eLTER_getNetworkEnvCharacts}
\usage{
getNetworkEnvCharacts(networkDEIMSID)
}
\arguments{
\item{networkDEIMSID}{is a DEIMS iD of network make from DEIMS-SDR website. More information about DEIMS iD in this page https://deims.org/docs/deimsid.html, and at this page https://deims.org/search?f%5B0%5D=result_type%3Anetwork the complete list of iLTER networks.}
}
\value{
The output of the function is a dataframe containing all the Environmental Characteristics of the network's sites.
}
\description{
This function ...
}
\examples{
listEnvCharacts <- getNetworkEnvCharacts(networkDEIMSID = 'https://deims.org/network/7fef6b73-e5cb-4cd2-b438-ed32eb1504b3')
knitr::kable(listEnvCharacts[1:5, ])

}
\author{
Alessandro Oggioni, phD (2020) <oggioni.a@irea.cnr.it>
}
