% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNetworkParameters.R
\name{getNetworkParameters}
\alias{getNetworkParameters}
\title{eLTER_getNetworkParameters}
\usage{
getNetworkParameters(networkDEIMSID)
}
\arguments{
\item{networkDEIMSID}{is a DEIMS iD of network make from DEIMS-SDR website. More information about DEIMS iD in this page https://deims.org/docs/deimsid.html, and at this page https://deims.org/search?f%5B0%5D=result_type%3Anetwork the complete list of iLTER networks.}
}
\value{
The output of the function is ...
}
\description{
This function ...
}
\examples{
listParams <- getNetworkParameters(networkDEIMSID = 'https://deims.org/network/7fef6b73-e5cb-4cd2-b438-ed32eb1504b3')
knitr::kable(listParams[1:10, ] \%>\% dplyr::rows_insert(tibble(parameterLabel = "...", parameterUri = "...")))

}
\author{
Alessandro Oggioni, phD (2020) <oggioni.a@irea.cnr.it>
}
