% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNetworkResearchTopics.R
\name{getNetworkResearchTopics}
\alias{getNetworkResearchTopics}
\title{eLTER_getNetworkResearchTopics}
\usage{
getNetworkResearchTopics(networkDEIMSID)
}
\value{
The output of the function is ...
}
\description{
This function ...
}
\examples{
getNetworkResearchTopics(networkDEIMSID = 'https://deims.org/network/7fef6b73-e5cb-4cd2-b438-ed32eb1504b3')

}
\author{
Alessandro Oggioni, phD (2020) <oggioniale@gmail.com>
}
