% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNetworkSites.R
\name{getNetworkSites}
\alias{getNetworkSites}
\title{eLTER_spatial_points_of_sites}
\usage{
getNetworkSites(networkDEIMSID)
}
\arguments{
\item{networkDEIMSID}{is a DEIMS iD of network make from DEIMS-SDR website. More information about DEIMS iD in this page https://deims.org/docs/deimsid.html, and at this page https://deims.org/search?f%5B0%5D=result_type%3Anetwork the complete list of iLTER networks.}
}
\value{
The output of the function is SpatialPointsDataFrame (package sp).
}
\description{
This function provide a SpatialPointsDataFrame (sp) of sites in a network provided by DEIMS-SDR.
}
\examples{
getNetworkSites(networkDEIMSID = 'https://deims.org/network/7fef6b73-e5cb-4cd2-b438-ed32eb1504b3')

}
\author{
Alessandro Oggioni, phD (2020) <oggioniale@gmail.com>
}
