% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSiteAffiliations.R
\name{getSiteAffiliations}
\alias{getSiteAffiliations}
\title{eLTER getSiteAffiliations function}
\usage{
getSiteAffiliations(deimsid)
}
\arguments{
\item{deimsid}{A character. The DEIMS ID of the site from
DEIMS-SDR website. More information about DEIMS ID in this pages:
\href{https://deims.org/docs/deimsid.html}{page}.}
}
\value{
The output of the function is a `tibble` with main features of the
site and the affiliations information, such as: networks and projects in
which the site is involved.
}
\description{
This function obtains details about an eLTER site
through the DEIMS-SDR sites API.
}
\examples{
tSiteAffiliation <- getSiteAffiliations(
  deimsid = "https://deims.org/f30007c4-8a6e-4f11-ab87-569db54638fe"
)
tSiteAffiliation

}
\author{
Alessandro Oggioni, phD (2020) \email{oggioni.a@irea.cnr.it}
}
