% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSiteContact.R
\name{getSiteContact}
\alias{getSiteContact}
\title{eLTER getSiteContact function}
\usage{
getSiteContact(deimsid)
}
\arguments{
\item{deimsid}{A `character`. It is the DEIMS ID of the site from
DEIMS-SDR website. More information about DEIMS ID from:
\href{https://deims.org/docs/deimsid.html}{page}.}
}
\value{
The output of the function is a `tibble` with main features of the
site and the contact information, such as: site manager, operation
organization, metadata provider, founding agency and site url.
}
\description{
This function obtains the contact information for an
eLTER site througth the DEIMS-SDR sites API.
}
\examples{
tSiteContact <- getSiteContact(
  deimsid = "https://deims.org/17210eba-d832-4759-89fa-9ff127cbdf6e"
)
tSiteContact

}
\author{
Alessandro Oggioni, phD (2020) \email{oggioni.a@irea.cnr.it}
}
