% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSiteGeneral.R
\name{getSiteGeneral}
\alias{getSiteGeneral}
\title{eLTER_getSiteGeneral}
\usage{
getSiteGeneral(deimsid)
}
\arguments{
\item{deimsid}{is a DEIMS iD of network make from DEIMS-SDR website. More information about DEIMS iD in this page https://deims.org/docs/deimsid.html.}
}
\value{
The output of the function is a tibble with main features of the site and the general information, such as: abstract, keywords, porpose, status, yearEstablished, yearClosed, hierarchy, siteName, short name, site type, protection level, images.
}
\description{
This function allows to obtain the general info of the eLTER site througth the DEIMS-SDR sites API.
}
\examples{
getSiteGeneral(deimsid = 'https://deims.org/17210eba-d832-4759-89fa-9ff127cbdf6e')

}
\author{
Alessandro Oggioni, phD (2020) <oggioni.a@irea.cnr.it>
}
