% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSiteInfrastructure.R
\name{getSiteInfrastructure}
\alias{getSiteInfrastructure}
\title{eLTER getSiteInfrastructure function}
\usage{
getSiteInfrastructure(deimsid)
}
\arguments{
\item{deimsid}{A `character`. The DEIMS ID of the site from
DEIMS-SDR website. More information about DEIMS ID from:
\href{https://deims.org/docs/deimsid.html}{page}.}
}
\value{
The output of the function is a `tibble` with main features of the
site and infrastructure information where available, such as:
power supply, accessibility, maintenaince interval, etc.
}
\description{
This function obtains infrastructure information 
of an eLTER site througth the DEIMS-SDR sites API.
}
\examples{
tSiteInfrastructure <- getSiteInfrastructure(
  deimsid = "https://deims.org/17210eba-d832-4759-89fa-9ff127cbdf6e"
)
tSiteInfrastructure

}
\author{
Alessandro Oggioni, phD (2020) \email{oggioni.a@irea.cnr.it}
}
