% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSiteRelatedResources.R
\name{getSiteRelatedResources}
\alias{getSiteRelatedResources}
\title{eLTER getSiteRelatedResources function}
\usage{
getSiteRelatedResources(deimsid)
}
\arguments{
\item{deimsid}{A `character`. The DEIMS ID of the site from
DEIMS-SDR website. More information about DEIMS ID from:
\href{https://deims.org/docs/deimsid.html}{page}.}
}
\value{
The output of the function is a `tibble` with main features of the
site and a list of the related resources collected by site.
}
\description{
This function obtains a list of related resources
collected in an eLTER site through the DEIMS-SDR sites API.
}
\examples{
tSiteRelatedResources <- getSiteRelatedResources(
  deimsid = "https://deims.org/f30007c4-8a6e-4f11-ab87-569db54638fe"
)
tSiteRelatedResources

}
\author{
Alessandro Oggioni, phD (2020) \email{oggioni.a@irea.cnr.it}
}
