% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSiteResearchTopics.R
\name{getSiteResearchTopics}
\alias{getSiteResearchTopics}
\title{eLTER_getSiteResearchTopics}
\usage{
getSiteResearchTopics(deimsid)
}
\value{
The output of the function is a tibble with main features of the site and the research topics envisaged into site.
}
\description{
This function allows to obtain the list of research topics of the eLTER site througth the DEIMS-SDR sites API.
}
\examples{
getSiteResearchTopics(deimsid = 'https://deims.org/f30007c4-8a6e-4f11-ab87-569db54638fe')

}
\author{
Alessandro Oggioni, phD (2020) <oggioni.a@irea.cnr.it>
}
