% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSitesByCountry.R
\name{getSitesByCountry}
\alias{getSitesByCountry}
\title{eLTER getSitesByCountry function}
\usage{
getSitesByCountry(country_name)
}
\arguments{
\item{country_name}{A `character`. Partial names are also matched.
However the partial string should be unique.
i.e. "United" can match "United States" and "United Kingdom"}
}
\value{
A `data.frame` of the sites within the given country
}
\description{
This function allows the user to retrieve a list of sites
associated with a country name. The user supplied country_name
must appear in the 'title' string of the site in DEIMS metadata
}
\examples{
sites_filtered <- getSitesByCountry("Austri")
(Austria, not Australia)
sites_filtered 
}
\author{
Micha Silver, \email{silverm@post.bgu.ac.il}
}
