% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSOSProcedureList.R
\name{getSOSProcedureList}
\alias{getSOSProcedureList}
\title{getSOSProcedureList}
\usage{
getSOSProcedureList(sosHost)
}
\arguments{
\item{sosHost}{a `character`. It is a SOS endpoint.}
}
\value{
The output of the function is a `list` with the name and URI (Uniform Resource Identifier) of each procedure.
}
\description{
This function list the procedures of a Sensor Observations Service (SOS).
}
\examples{
getSOSProcedureList(sosHost = 'http://getit.lteritalia.it')

}
\author{
Alessandro Oggioni, phD (2020) <oggioni.a@irea.cnr.it>
}
