% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_location_info.R
\name{map_add_location}
\alias{map_add_location}
\title{Create location map}
\usage{
map_add_location(location_data, map = NULL)
}
\arguments{
\item{location_data}{A \code{sf} object. It is the location geodata provided by
\code{get_location_info()}.}

\item{map}{A \code{leaflet} object. It is an empty map by default.}
}
\description{
This is an internal function for create only the map of
location.
}
\examples{
\dontrun{
## Not run:
map <- map_add_location(
  location_data = location$data,
  map = NULL
)
}
## End (Not run)

}
\author{
Alessandro Oggioni, phD (2020) \email{oggioni.a@irea.cnr.it}

Paolo Tagliolato, phD (2020) \email{tagliolato.p@irea.cnr.it}
}
\keyword{internal}
