% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eLTER_reporting_occ.R
\name{map_occ_inat2elter}
\alias{map_occ_inat2elter}
\title{Harmonize outputs of get_site_speciesOccurrence and map them into eLTER
reporting format}
\usage{
map_occ_inat2elter(x, deimsid, version = "1.3")
}
\arguments{
\item{x}{A \code{tibble} like one that can be obtained by
as_tibble(get_site_speciesOccurrences(deimsid, "gbif")$gbif)}

\item{deimsid}{A \code{character}. The DEIMS.iD of the site from
DEIMS-SDR website. DEIMS ID information}

\item{version}{A \code{character} for select the version of eLTER
Data Reporting Format. Default 1.3}
}
\value{
list with the following named elements:
\itemize{
\item deimsid: the same deimsid passed in input
\item source: one of "gbif", "inat", "obis"
\item data_mapping: tibble structured according to data_mapping of
eLTER reporting format
\item reference_TAXA: tibble structured according to reference_TAXA of
eLTER reporting format
\item reference_VARIABLES: tibble structured according to reference_VARIABLES
of eLTER reporting format
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
}
\author{
Paolo Tagliolato, phD \email{tagliolato.p@irea.cnr.it}

Martina Zilioli \email{zilioli.m@irea.cnr.it}

Alessandro Oggioni, phD \email{oggioni.a@irea.cnr.it}
}
