% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\docType{data}
\name{package_settings}
\alias{package_settings}
\title{Package settings that can be changed by the user}
\format{
An object of class \code{environment} of length 1.
}
\usage{
package_settings
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
\seealso{
Other package_customizable_settings: 
\code{\link{get_deims_base_url}()}
}
\concept{package_customizable_settings}
\keyword{datasets}
