% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parametersChartWaffle.R
\name{parameterChartsWaffle}
\alias{parameterChartsWaffle}
\title{Waffle chart of parameter collected in the site/network}
\usage{
parameterChartsWaffle(deimsid)
}
\arguments{
\item{deimsid}{is a DEIMS iD of network make from DEIMS-SDR website. More information about DEIMS iD in this page https://deims.org/docs/deimsid.html.}
}
\value{
The output of the function is a waffle chart.
}
\description{
This function allows to obtain a waffle chart of the parameter collected in a site or network grouped into compounds.
}
\examples{
parameterChartsWaffle(deimsid = 'https://deims.org/f30007c4-8a6e-4f11-ab87-569db54638fe')

}
\author{
Alessandro Oggioni, phD (2020) <oggioni.a@irea.cnr.it>
}
