% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_site_MODIS.R
\name{plot_agg_map}
\alias{plot_agg_map}
\title{Aggregate time series of MODIS images
Save and show a map}
\usage{
plot_agg_map(product, output_dir, site_name, agg_function = "mean")
}
\arguments{
\item{output_dir}{\code{character}, where MODIS images were saved
This directory is returned by \code{get_site_MODIS()}
The final map, as png image file will be saved here also.}
}
\value{
Full paths to saved Geotiff rasters
}
\description{
Aggregate time series of MODIS images
Save and show a map
}
\details{
Read all vrt time series images in \code{output_dir}
Prepare an aggregation raster of all maps in the time series
Save and show a plot of the aggregated map

This function is not exported. It is called by \code{get_site_MODIS()}
}
\examples{
 \dontrun{
}

}
\references{
\insertRef{sfR}{ReLTER}

\insertRef{terraR}{ReLTER}
}
\author{
Micha Silver, phD (2020) \email{silverm@post.bgu.ac.il}

Alessandro Oggioni, phD (2020) \email{oggioni.a@irea.cnr.it}
}
