% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_site_MODIS.R
\name{plot_timeseries}
\alias{plot_timeseries}
\title{Plot a time series of average pixel values from MODIS images
cropped to an eLTER site boundary.}
\usage{
plot_timeseries(deimsid, dataset, output_dir)
}
\arguments{
\item{dataset}{A \code{character}. The requested dataset. One of:
"LST_day", "LST_night", VI"
Default is "VI".}

\item{output_dir}{a \code{character}, where MODIS images were saved
This directory is returned by \code{get_site_MODIS()}
The final graph as png image file will be saved here also.}
}
\value{
Full path to the saved png image.
}
\description{
\href{https://deims.org/docs/deimsid.html}{here}.
#' @param deimsid  A \code{character}. The DEIMS ID of the site from
DEIMS-SDR website. DEIMS ID information
\href{https://deims.org/docs/deimsid.html}{here}.
}
\details{
Read all images in \code{output_dir} and prepare line plots
of average values over the site boundary for each band.
}
\examples{
 \dontrun{

}

}
\references{
\insertRef{sfR}{ReLTER}

\insertRef{terraR}{ReLTER}
}
\author{
Micha Silver, phD (2020) \email{silverm@post.bgu.ac.il}

Alessandro Oggioni, phD (2020) \email{oggioni.a@irea.cnr.it}
}
