% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produce_site_parameters_waffle.R
\name{produce_site_parameters_waffle}
\alias{produce_site_parameters_waffle}
\title{eLTER produce_site_parameters_waffle function}
\usage{
produce_site_parameters_waffle(deimsid)
}
\arguments{
\item{deimsid}{A `character`. The DEIMS ID of site/network from:
DEIMS-SDR website. More information about DEIMS ID from:
\href{https://deims.org/docs/deimsid.html}{page}.}
}
\value{
The output of the function is a waffle chart.
}
\description{
This function produces a waffle chart of the parameters
collected in a site or network grouped into compounds.
}
\examples{
\dontrun{
waffle <- produce_site_parameters_waffle(
  deimsid = "https://deims.org/f30007c4-8a6e-4f11-ab87-569db54638fe"
)
print(waffle)
}

}
\author{
Alessandro Oggioni, phD (2020) \email{oggioni.a@irea.cnr.it}
}
