% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produce_site_qrcode.R
\name{produce_site_qrcode}
\alias{produce_site_qrcode}
\title{eLTER produce_site_qrcode function}
\usage{
produce_site_qrcode(deimsid)
}
\arguments{
\item{deimsid}{A `character`. The DEIMS ID of site from
DEIMS-SDR website. More information about DEIMS iD in this
\href{https://deims.org/docs/deimsid.html}{page}.}
}
\value{
The output of the function is a QRcode `image`.
}
\description{
This function obtain the QRCode of any DEIMS-SDR
entities of the eLTER.
}
\examples{
produce_site_qrcode(
  deimsid = "https://deims.org/f30007c4-8a6e-4f11-ab87-569db54638fe"
)

}
\author{
Alessandro Oggioni, phD (2020) \email{oggioni.a@irea.cnr.it}
}
