% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produce_zenodo_record.R
\name{produce_zenodo_record}
\alias{produce_zenodo_record}
\title{Upload dataset to Zenodo}
\usage{
produce_zenodo_record(
  mytoken,
  myfiles,
  delim,
  elter_data_format = TRUE,
  upload_rdata = TRUE,
  dataset_type = "dataset",
  dataset_title,
  dataset_description,
  dataset_authors,
  dataset_license = "CC-BY-SA-4.0",
  dataset_accessRight = "open",
  dataset_version = "1.0",
  dataset_language = "eng",
  dataset_keywords,
  dataset_relatedIdentifier,
  dataset_communities = c("lter-italy", "elter"),
  dataset_grants,
  dataset_publish = FALSE
)
}
\arguments{
\item{mytoken}{A `character`. Scopes assign permissions to your personal
access token. A personal access token works just like a normal OAuth access
token for authentication against the API. This token can be created at
application
\href{https://zenodo.org/account/settings/applications/tokens/new/}{page}.}

\item{myfiles}{A `character`. The list of the files to deposit in Zenodo.
Please provide all files only with 'csv' extension.}

\item{delim}{A `character`. Provide the character used to separate fields
within a record. Only if the extension of the file(s) are 'csv', 'tsv', or
'txt'.}

\item{elter_data_format}{A `logical`. Is the dataset in the elter data
reporting format? Default TRUE.}

\item{upload_rdata}{A `logical`. Do you  want also upload a version of
the dataset in RData format onto this record?
Choosing this option you will have a record improvement ...
Default TRUE.}

\item{dataset_title}{A `character`. The title of the dataset.}

\item{dataset_description}{A `character`. The description of the dataset.}

\item{dataset_authors}{A `tibble`. It is a list of creator(s) of the
dataset. The approach is to use the firstname, lastname, affiliation,
orcid of the authors of the dataset. Please follow the example.}

\item{dataset_license}{A `character`. It is the license with which the
dataset is released. The license should be set with the Zenodo id of the
license. Default "CC-BY-SA-4.0".
The supported licenses values are from \url{https://opendefinition.org}
and \url{https://spdx.org}.}

\item{dataset_accessRight}{A `character`. Default "open". Other options are:
'embargoed', 'restricted' and 'closed'.}

\item{dataset_version}{A `character`. It is a version of the dataset.
Default "1.0".}

\item{dataset_language}{A `character`. It is the language of the dataset.
Only one value is possible. Default "eng".}

\item{dataset_keywords}{A `character`. A multiple values are possible. The
keyword to the record of dataset.}

\item{dataset_relatedIdentifier}{A `tibble`. It is the related entities
(e.g. scientific paper, data paper, etc.) with the detaset. The tibble is
composed by 2 variables: relation and identifier. Relation can be one of
among following values: "isCitedBy", "cites", "isSupplementTo",
"isSupplementedBy", "isNewVersionOf", "isPreviousVersionOf",
"isPartOf", "hasPart", "compiles", "isCompiledBy", "isIdenticalTo",
"isAlternateIdentifier". While identifier is any type of identifier
(e.g. DOI).}

\item{dataset_grants}{A `character`. A multiple values are possible. Put a
list of project identifier as well as showed by the European Commission via
\href{http://www.openaire.eu/}{OpenAIRE}.}

\item{dataset_publish}{A `logical`. State whether the Zenodo record is to
be published. The parameter publish can be set to TRUE (to use CAUTIOUSLY,
only if you want to publish your record). Default FALSE.}

\item{dataset_community}{A `character`. It is a name of community as created
in Zenodo. A multiple values are possible. Default "lter-italy" and "elter".}
}
\value{
A link (URL) of the deposited record in Zenodo. The user must then:
1. visit the webpage of record, 2. check the information provided, 3.
'save' and 'publish' the dataset on Zenodo repository, 4. use the DOI to
cite the record.
}
\description{
`r lifecycle::badge("experimental")`
This function allows to deposit a dataset (only one dataset at the moment)
to \href{https://zenodo.org}{Zenodo} repository. The function use the 
facilities of `zen4r` package. Blondel, Emmanuel, & Barde, Julien. (2021).
zen4R: R Interface to Zenodo REST API (0.5-2). Zenodo.
\url{https://doi.org/10.5281/zenodo.5741143}.
}
\examples{
\dontrun{
## Not run:

authors <- tibble::tibble(
  name = c("Luke", "Leia"),
  surname = c("Skywalker", "Organa"),
  affiliation = c("Tatooine", "Alderaan"),
  orcid = c("0000-0002-7997-219X", "0000-0002-7997-219X")
)
keywords <- c("Star Wars", "species", "films", "planets")
relatedIdentifiers <- tibble::tibble(
  relation = c("isSupplementTo", "isPartOf"),
  identifier = c("10.1038/s4150-01-0032", "10.1016/j.2051.06.026")
)
grants <- c("871128", "654359", "871126")
produce_zenodo_record(
  mytoken = 'HOwWaUAel6C4j5UkOgWlLGCSkAkkiKuTCjlqJcfQ16OvGCjIL9SrLQBC1hAU',
  myfiles = c(
   "miscellaneus/file_show/data_mapping.csv",
   "miscellaneus/file_show/reference_TAXA.csv",
   "miscellaneus/file_show/reference_VARIABLES.csv"
  ),
  delim = ";",
  elter_data_format = TRUE,
  upload_rdata = TRUE,
  dataset_title = "The title",
  dataset_description = "This is the description of the dataset.",
  dataset_authors = authors,
  dataset_keywords = keywords,
  dataset_relatedIdentifier = relatedIdentifiers,
  dataset_communities = "lter-italy",
  dataset_grants = grants,
  dataset_publish = FALSE
)
}

## End (Not run)

}
\author{
Alessandro Oggioni, phD (2020) \email{oggioni.a@irea.cnr.it}
}
