% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produce_zenodo_record.R
\name{produce_zenodo_record_from_elter_reporting}
\alias{produce_zenodo_record_from_elter_reporting}
\title{Create a record in Zenodo from eLTER data reporting format}
\usage{
produce_zenodo_record_from_elter_reporting(
  x,
  saveRDS = FALSE,
  filepath = tempdir(),
  filename,
  mytoken,
  record_title,
  record_description,
  record_authors,
  record_license = "CC-BY-SA-4.0",
  record_accessRight = "open",
  record_version = "1.0",
  record_language = "eng",
  record_keywords,
  record_relatedIdentifier,
  record_communities = c("lter-italy", "elter"),
  record_grants
)
}
\arguments{
\item{x}{A \code{list} like the one created by function
\code{reporting_produce_data_object_v1.3}}

\item{saveRDS}{A \code{logical}. Save also object in RDS format.
Defaults to FALSE.}

\item{filepath}{A \code{character} file path. Defaults to temporary directory}

\item{filename}{A \code{character}. Optional filename associated with the object,
of the form provided as output by the function
\code{reporting_compose_file_name}.
Defaults to random string}

\item{mytoken}{A \code{character}. Scopes assign permissions to your personal
access token. A personal access token works just like a normal OAuth access
token for authentication against the API. This token can be created at
application
\href{https://zenodo.org/account/settings/applications/tokens/new/}{page}.}

\item{record_title}{A \code{character}. The title of the record.}

\item{record_description}{A \code{character}. The description of the record.}

\item{record_authors}{A \code{tibble}. It is a list of creator(s) of the
record. The approach is to use the firstname, lastname, affiliation,
orcid of the authors of the record. Please follow the example.}

\item{record_license}{A \code{character}. It is the license with which the
record is released. The license should be set with the Zenodo id of the
license. Default "CC-BY-SA-4.0".
The supported licenses values are from \url{https://opendefinition.org}
and \url{https://spdx.org}.}

\item{record_accessRight}{A \code{character}. Default "open". Other options are:
'embargoed', 'restricted' and 'closed'.}

\item{record_version}{A \code{character}. It is a version of the record.
Default "1.0".}

\item{record_language}{A \code{character}. It is the language of the record.
Only one value is possible. Default "eng".}

\item{record_keywords}{A \code{character}. A multiple values are possible. The
keyword to the record of record.}

\item{record_relatedIdentifier}{A \code{tibble}. It is the related entities
(e.g. scientific paper, data paper, etc.) with the detaset. The tibble is
composed by 2 variables: relation and identifier. Relation can be one of
among following values: "isCitedBy", "cites", "isSupplementTo",
"isSupplementedBy", "isNewVersionOf", "isPreviousVersionOf",
"isPartOf", "hasPart", "compiles", "isCompiledBy", "isIdenticalTo",
"isAlternateIdentifier". While identifier is any type of identifier
(e.g. DOI).}

\item{record_communities}{A \code{character}. It is a name of communities as
created in Zenodo. A multiple values are possible. Default "lter-italy"
and "elter".}

\item{record_grants}{A \code{character}. A multiple values are possible. Put a
list of project identifier as well as showed by the European Commission via
\href{http://www.openaire.eu/}{OpenAIRE}.}
}
\value{
A link (URL) of the deposited record in Zenodo. The user must then:
\enumerate{
\item visit the webpage of record, 2. check the information provided, 3.
'save' and 'publish' the record on Zenodo repository, 4. use the DOI to
cite the record.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function allows to deposit a record to
\href{https://zenodo.org}{Zenodo} repository with the eLTER data reporting
format.
The function use the functions implemented by \code{zen4r} package.
Blondel, Emmanuel, & Barde, Julien. (2021).
zen4R: R Interface to Zenodo REST API (0.5-2). Zenodo.
\url{https://doi.org/10.5281/zenodo.5741143}.
}
\examples{
\dontrun{
## Not run:

deimsid <- "https://deims.org/8eda49e9-1f4e-4f3e-b58e-e0bb25dc32a6"
time_span <- 2015 # e.g. whole year
time_span <- "20150302-20180415" # e.g. span between two dates
data_topic <- "VEG" # data provider defined abbreviation of "vegetation"
variable_group <- "SPECCOVER" # data provider defined abbreviation
version <- "V20220907"

filename <- reporting_compose_file_name(
  deimsid = deimsid,
  data_topic = data_topic,
  variable_group = variable_group,
  time_span = time_span,
  version = version
)

data <- dplyr::tribble(
  ~id, ~value, 
  1, 7.5, 
  2, 4.2
)
station <- dplyr::tribble(
  ~SITE_CODE, ~STATION_CODE, ~STYPE, ~LAT,      ~LON,       ~ALTITUDE,
  deimsid,    "IP2",         "AREA",  45.340805, 7.88887495, 265
)
method <- dplyr::tribble(
  ~VARIABLE, ~METH_DESCR,
  "COVE_F",  "Analysis of ammonium..."  
)
  
research_object <- reporting_produce_data_object_v1.3(
 filename = filename,
 deimsid = deimsid,
 data = data,
 station = station,
 method = method
)

authors <- tibble::tibble(
  name = c("Luke", "Leia"),
  surname = c("Skywalker", "Organa"),
  affiliation = c("Tatooine", "Alderaan"),
  orcid = c("0000-0002-7997-219X", "0000-0002-7997-219X")
)
keywords <- c("Star Wars", "species", "films", "planets")
relatedIdentifiers <- tibble::tibble(
  relation = c("isSupplementTo", "isPartOf"),
  identifier = c("10.1038/s4150-01-0032", "10.1016/j.2051.06.026")
)
grants <- c("871128", "654359", "871126")

produce_zenodo_record_from_elter_reporting(
  x = research_object,
  saveRDS = TRUE,
  filepath = ".",
  filename = filename,
  mytoken = mytoken, # generate your Zenodo token
  record_title = "The title of eLTER reporting",
  record_description = "This is the description of the record of eLTER data
    reporting format.",
  record_authors = authors,
  record_license = "CC-BY-SA-4.0",
  record_accessRight = "open",
  record_version = "1.0",
  record_language = "eng",
  record_keywords = keywords,
  record_relatedIdentifier = relatedIdentifiers,
  record_communities = "lter-italy",
  record_grants = grants
)
}

## End (Not run)

}
\seealso{
zen4R documentation \url{https://github.com/eblondel/zen4R/wiki}
}
\author{
Alessandro Oggioni, phD \email{oggioni.a@irea.cnr.it}

Paolo Tagliolato, phD \email{tagliolato.p@irea.cnr.it}
}
