% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ReLTER-package.R
\docType{package}
\name{ReLTER-package}
\alias{ReLTER}
\alias{ReLTER-package}
\title{ReLTER: An Interface for the eLTER Community}
\description{
\if{html}{\figure{logo.png}{options: style='float: right' alt='logo' width='120'}}

ReLTER provides access to DEIMS-SDR (https://deims.org/), and allows interaction with data and software implemented by eLTER Research Infrastructure (RI) thus improving data sharing among European LTER projects. ReLTER uses the R language to access and interact with the DEIMS-SDR archive of information shared by the Long Term Ecological Research (LTER) network. This package grew within eLTER H2020 as a major project that will help advance the development of European Long-Term Ecosystem Research Infrastructures (eLTER RI - https://elter-ri.eu). The ReLTER package functions in particular allow to: - retrieve the information about entities (e.g. sites, datasets, and activities) shared by DEIMS-SDR (see e.g. get_site_info function); - interact with the ODSEurope (maps.opendatascience.eu) starting with the dataset shared by DEIMS-SDR (https://deims.org/) (see e.g. the get_site_ODS() function); - use the eLTER site informations to download and crop geospatial data from other platforms (see e.g. get_site_ODS function()); - improve the quality of the dataset (see e.g. get_id_worms()). Functions currently implemented are derived from discussions of the needs among the eLTER users community. The ReLTER package will continue to follow the progress of eLTER-RI and evolve, adding new tools and improvements as required.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://docs.ropensci.org/ReLTER}
  \item Report bugs at \url{https://github.com/ropensci/ReLTER/issues}
}

}
\author{
\strong{Maintainer}: Alessandro Oggioni \email{alessandro.oggioni@cnr.it} (\href{https://orcid.org/0000-0002-7997-219X}{ORCID}) [funder]

Authors:
\itemize{
  \item Alessandro Oggioni \email{alessandro.oggioni@cnr.it} (\href{https://orcid.org/0000-0002-7997-219X}{ORCID}) [funder]
  \item Micha Silver \email{silverm@post.bgu.ac.il} (\href{https://orcid.org/0000-0002-1128-1325}{ORCID}) [contributor]
  \item Paolo Tagliolato \email{tagliolato.a@irea.cnr.it} (\href{https://orcid.org/0000-0002-0261-313X}{ORCID}) [contributor]
  \item Luigi Ranghetti \email{luigi@ranghetti.info} (\href{https://orcid.org/0000-0001-6207-5188}{ORCID}) [contributor]
}

Other contributors:
\itemize{
  \item Allison Horst (Allison reviewed the package (v. 1.0.0) for rOpenSci, see https://github.com/ropensci/software-review/issues/485) [reviewer]
  \item Will Bolton (Will reviewed the package (v. 1.0.0) for rOpenSci, see https://github.com/ropensci/software-review/issues/485) [reviewer]
  \item Mauro Lepore (Mauro was editor for rOpenSci, see https://github.com/ropensci/software-review/issues/485) [editor]
}

}
\keyword{internal}
