% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{set_deims_base_url}
\alias{set_deims_base_url}
\title{Set DEIMS-SDR API base URL}
\usage{
set_deims_base_url(url = "https://deims.org/", force = FALSE)
}
\arguments{
\item{url}{A \code{character}. Set the base URL to DEIMS-SDR.}

\item{force}{A \code{boolean}. Default FALSE.}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
}
