% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{setDeimsBaseUrl}
\alias{setDeimsBaseUrl}
\title{set deims api base url}
\usage{
setDeimsBaseUrl(url = "https://deims.org/", force = FALSE)
}
\arguments{
\item{url}{A \code{character}. Set the base url to deims.}
}
\description{
set deims api base url
}
