% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_id_gbif.R
\name{taxon_id_gbif}
\alias{taxon_id_gbif}
\title{Enrich and certify a list of species names by
comparing with \href{https://www.gbif.org}{GBIF}.}
\usage{
taxon_id_gbif(input, taxaColumn = 1, refine = FALSE)
}
\arguments{
\item{input}{A \code{tibble}. The table that contain the species
names list to be checked.}

\item{taxaColumn}{A \code{numeric}. The cardinal number of the column where
species list is. Default is \code{1}.}

\item{refine}{A \code{logical}. With this selection, the function allows to
refine the result(s) that match with more GBIF records. By a interactive
use of the terminal, the user can chose the result. Default is \code{FALSE}.}
}
\value{
When the \code{refine} is \code{FALSE} the output of the function is a
\code{tibble} with the refinement of the input, else the function simply
returns the unchanged input.

All the labels of the columns are the terms of
\href{https://dwc.tdwg.org/terms/#record-level}{Darwin Core terms}.
The columns labels are annotate with the link (URI) of the
\href{https://dwc.tdwg.org/terms/#record-level}{Darwin Core terms}
as attributes of the \code{tibble}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
This function provide tibble object with all the columns of input table
of taxa plus new columns obtained from:
\href{https://data-blog.gbif.org/post/gbif-backbone-taxonomy/}{
GBIF Backbone Taxonomy}.
This function use the \code{rgbif::name_backbone_checklist} function for
interact with API backbone taxonomy of GBIF.
}
\examples{
macrophytes <- tibble::tibble(
  ID = c(1:18),
  species = c("Ceratophyllum demersum",
  "Lagarosiphon major",
  "Littorella uniflora",
  "Myriophyllum spicatum",
  "Myriophyllum verticillatum",
  "Najas marina",
  "Najas minor",
  "Nelumbo nucifera",
  "Nitella gracilis",
  "Nuphar lutea",
  "Nymphaea alba",
  "Nymphoides peltata",
  "Persicaria amphibia",
  "Potamogeton perfoliatus",
  "Potamogeton pusillus",
  "Ranunculus sp.",
  "Stuckenia pectinata",
  "Trapa natans")
)
table <- taxon_id_gbif(
  input = macrophytes,
  taxaColumn = 2,
  refine = TRUE
)
table

# The annotated URIs of columns label are achieved by:
attributes(table)$uri

}
\references{
\insertRef{gbifR}{ReLTER}

\insertRef{dplyrR}{ReLTER}

\insertRef{tibbleR}{ReLTER}
}
\author{
Alessandro Oggioni, phD (2021) \email{oggioni.a@irea.cnr.it}
}
