% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_id_worms.R
\name{taxon_id_worms}
\alias{taxon_id_worms}
\title{Enrich and certify a list of species names by
comparing with \href{https://www.marinespecies.org}{WoRMS}.}
\usage{
taxon_id_worms(input, taxaColumn = 1, verbose = TRUE, refine = FALSE)
}
\arguments{
\item{input}{A \code{tibble}. The table that contain the species
names list to be checked.}

\item{taxaColumn}{A \code{numeric}. The cardinal number of the column where
species list is. Default is \code{1}.}

\item{verbose}{A \code{logical}. Whit this selection, the function returns a
message with number of record(s) that don't match with any Worms names and
the number of record(s) that match with more that one Worms name. Default
is \code{TRUE}.}

\item{refine}{A \code{logical}. With this selection, the function allows to
refine the result(s) that match with more Worms records. By a interactive
use of the terminal, the user can chose the result. Default is \code{FALSE}.}
}
\value{
The output of the function is a \code{tibble} with the columns provided
and new columns such as: valid_name, valid_authority, valid_AphiaID, status,
synonyms, LSID, url, matchType, nOfWormsRecords, wormsRecords obtained by
\href{http://www.marinespecies.org/rest/}{Worms rest API}.
The function also return, if verbose is TRUE, the list of records that don't
match with Worms name species.

Most of the labels of the columns are the terms of
\href{https://dwc.tdwg.org/terms/#record-level}{Darwin Core terms}.
The columns labels are annotate with the link (URI) of the
\href{https://dwc.tdwg.org/terms/#record-level}{Darwin Core terms}
as attributes of the \code{tibble}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
This function provide tibble object with all the columns of input table
of taxa plus new columns such as valid_name, valid_authority, valid_AphiaID,
status, synonyms, LSID, url, matchType, nOfWormsRecords, wormsRecords
obtained from Word Register of Marine Species
\href{http://www.marinespecies.org/rest/}{WoRMS rest API}.
}
\examples{
phytoplankton <- tibble::tibble(
   ID = c(1, 2, 3, 4, 5, 6, 7),
   species = c(
   "Asterionella formosa", "Chrysococcus sp.",
   "Cryptomonas rostrata", "Dinobryon divergens",
   "Mallomonas akrokomos", "Melosira varians",
   "Cryptomonas rostrata"
 )
)
table <- taxon_id_worms(
  input = phytoplankton,
  taxaColumn = 2,
  verbose = TRUE,
  refine = TRUE
)
table

# The annotated URIs of columns label are achieved by:
attributes(table)$uri

}
\seealso{
\code{\link[worrms:wm_records_names]{worrms::wm_records_names()}}
}
\author{
Alessandro Oggioni, phD (2021) \email{oggioni.a@irea.cnr.it}

Paolo Tagliolato, phD (2021) \email{tagliolato.p@irea.cnr.it}
}
