% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_id_worms_refine.R
\name{taxon_id_worms_refine}
\alias{taxon_id_worms_refine}
\title{eLTER taxon_id_worms_refine}
\usage{
taxon_id_worms_refine(
  input,
  taxaColumn = 1,
  interaction = FALSE,
  choiceNumber = NULL
)
}
\arguments{
\item{input}{a \code{tibble}. The output of taxon_ID_Worms.}

\item{taxaColumn}{a \code{numeric}. The cardinal number of the column where
species list is. Default is \code{1}.}

\item{interaction}{a \code{boolean}. When \code{TRUE} the user can interact with
the choice of one of the proposed matching. Default is \code{FALSE}.}

\item{choiceNumber}{a \code{numeric}. The number of selected choice. Default
is \code{NULL}. When \code{NULL} no choice is performed.}
}
\value{
In interactive mode or when the \code{choiceNumber} is \code{NULL} the
output of the function is a \code{tibble} with the refinement of the
input, else the function simply returns the unchanged input.
}
\description{
This function provides a refining of the output of
function taxon_id_worms. It prints a list of possible matches
and possibly enables the user to interactively make his choice.
}
\references{
\insertRef{dplyrR}{ReLTER}
}
\author{
Alessandro Oggioni, phD (2021) \email{oggioni.a@irea.cnr.it}

Paolo Tagliolato, PhD (2021) \email{tagliolato.p@irea.cnr.it}
}
\keyword{internal}
