% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_Q.R
\name{do_Q}
\alias{do_Q}
\title{eLTER do_Q function}
\usage{
do_Q(q, jj)
}
\arguments{
\item{q}{a character}

\item{jj}{a character}
}
\value{
The output of the function is ...
}
\description{
This function parses json content returned from an HTTP request
}
\author{
Paolo Tagliolato, PhD (2021) \email{tagliolato.p@irea.cnr.it}
}
\keyword{internal}
