% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/do_Q.R
\name{do_Q}
\alias{do_Q}
\title{do_Q}
\usage{
do_Q(q, jj)
}
\arguments{
\item{q}{a character}

\item{jj}{a character}
}
\value{
The output of the function is ...
}
\description{
This function ...
}
\examples{
require('dplyr')
deimsid <- 'https://deims.org/17210eba-d832-4759-89fa-9ff127cbdf6e'
url <- paste0("https://deims.org/", "api/sites/", substring(deimsid, 19))
export <- httr::GET(url = url)
jj <- httr::content(export, "text")

q = '{title: .title,
    uri: "\\\\(.id.prefix)\\\\(.id.suffix)",
    geoBoundaries: .attributes.geographic.boundaries,
    geoCoord: .attributes.geographic.coordinates,
    country: .attributes.geographic.country,
    geoElev: .attributes.geographic.elevation,
    envCharacteristics: .attributes.environmentalCharacteristics
  }'

do_Q(q, jj)
}
\author{
Paolo Tagliolato, phD (2021) \email{tagliolato.p@irea.cnr.it}
}
