% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_site_EcoDataCube.R
\name{EDC_construct_full_url}
\alias{EDC_construct_full_url}
\title{Construct full url for download}
\usage{
EDC_construct_full_url(edc_row, dataset_year, dataset_month)
}
\arguments{
\item{edc_row}{\code{vector} one row for chosen dataset from edc_df.}

\item{dataset_year}{\code{string} Chosen year}

\item{dataset_month}{\code{string} Chosen month}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Construct full URL, including replacing year and month where required.
}
\author{
Micha Silver, phD (2020) \email{silverm@post.bgu.ac.il}
}
