% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_site_EcoDataCube.R
\name{EDC_construct_full_url}
\alias{EDC_construct_full_url}
\title{Construct full url for download}
\usage{
EDC_construct_full_url(edc_row, from_yr, from_mon)
}
\arguments{
\item{edc_row}{Vector, one row for chosen dataset from edc_df.}

\item{from_yr}{Character Chosen year}

\item{from_mon}{Character Chosen month}
}
\description{
Construct full URL, including replaceing year and month where required.
}
\author{
Micha Silver, phD (2020) \email{silverm@post.bgu.ac.il}
}
