% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_activity_info.R
\name{get_activity_info}
\alias{get_activity_info}
\title{Return a tibble object containing information, as a stored in
\href{https://deims.org/}{DEIMS-SDR catalogue}, about a single eLTER
activity (e.g.
\url{https://deims.org/activity/8786fc6d-5d70-495c-b901-42f480182845}).}
\usage{
get_activity_info(activityid)
}
\arguments{
\item{activityid}{A `character`. It is the DEIMS.iD of activity make from
DEIMS-SDR website. DEIMS.iD information
\href{https://deims.org/docs/deimsid.html}{here}.
The DEIMS.iD of activity is the URL for the activity page.}
}
\value{
The output of the function is a `tibble` with main features of
the activities in a site, and a `leaflet` map plot.
}
\description{
This function obtains the information about of an eLTER
activity provided in DEIMS-SDR.
}
\examples{
activities <- get_activity_info(
  activityid =
  "https://deims.org/activity/8786fc6d-5d70-495c-b901-42f480182845"
)
activities

}
\author{
Alessandro Oggioni, phD (2020) \email{oggioni.a@irea.cnr.it}
}
