% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_dataset_info.R
\name{get_dataset_info}
\alias{get_dataset_info}
\title{Obtain the information about of an eLTER dataset.}
\usage{
get_dataset_info(datasetid)
}
\arguments{
\item{datasetid}{A `character`. It is the DEIMS ID of dataset make from
DEIMS-SDR website. DEIMS ID information 
\href{https://deims.org/docs/deimsid.html}{here}.
The DEIMS ID of dataset is the URL for the dataset page.}
}
\value{
The output of the function is a `tibble` with main features
of the site and the related resources collected by site.
}
\description{
This function obtains the information about of an eLTER
dataset (e.g.
\url{https://deims.org/activity/8786fc6d-5d70-495c-b901-42f480182845})
provided in \href{https://deims.org/}{DEIMS-SDR catalogue}.
}
\examples{
tDataset <- get_dataset_info(
  datasetid =
  "https://deims.org/dataset/38d604ef-decb-4d67-8ac3-cc843d10d3ef"
)
tDataset

}
\author{
Alessandro Oggioni, phD (2020) \email{oggioni.a@irea.cnr.it}
}
