% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_deims_api_version.R
\name{get_deims_API_version}
\alias{get_deims_API_version}
\title{get version of DEIMS api}
\usage{
get_deims_API_version(deims_url = getDeimsBaseUrl())
}
\arguments{
\item{deims_url}{A \code{character}. DEIMS base url. Defaults to package
settings.}
}
\value{
version number.
}
\description{
get version of DEIMS api
}
