% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/globals.R
\name{get_deims_base_url}
\alias{get_deims_base_url}
\title{Get DEIMS-SDR base URL}
\usage{
get_deims_base_url()
}
\value{
DEIMS-SDR base URL
}
\description{
Get DEIMS-SDR base URL
}
\seealso{
Other package_customizable_settings: 
\code{\link{package_settings}}
}
\concept{package_customizable_settings}
