% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_id.R
\name{get_id}
\alias{get_id}
\title{eLTER get_id function}
\usage{
get_id(deimsid, resource = "sites", test, ...)
}
\arguments{
\item{deimsid}{A `character`. It is the DEIMS.iD of the site, activity or
dataset from DEIMS-SDR website. DEIMS.iD information 
\href{https://deims.org/docs/deimsid.html}{here}.}

\item{resource}{Character: one among `"sites"` (default), `"activities"` or
`"datasets"` (`"networks"` currently not tested).}

\item{test}{Logical: if TRUE, content is not retrieved from
\url{https://deims.org} but internally (only for some defined IDs).
Used for testing purposes to avoid errors in case or HTTP error 500.
If missing, the content of the internal variable}

\item{...}{Arguments to be passed to `httr::RETRY()`.}

\item{jj}{a character}
}
\value{
A character containing the json content, and an attribute `status`.
}
\description{
Internal function to retrieve json content from ID
}
\author{
Alessandro Oggioni, phD (2020) \email{oggioni.a@irea.cnr.it}

Luigi Ranghetti, phD (2021) \email{luigi@ranghetti.info}
}
\keyword{internal}
