% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ilter_envcharacts.R
\name{get_ilter_envcharacts}
\alias{get_ilter_envcharacts}
\title{Obtain a list of all Environmental Characteristics of ILTER sites.}
\usage{
get_ilter_envcharacts(sitesNum = 0)
}
\arguments{
\item{sitesNum}{A `integer`. It is the number of the sites that are
read to get the information. Use this parameters moreover for provide
example of this function.}
}
\value{
The output of the function is a `tibble` containing all the
Environmental Characteristics of ILTER's sites.
}
\description{
This function obtains all Environmental Characteristics:
title, URI, geo-coordinates, country name, and elevation
of all
\href{https://www.ilter.network/network/global-coverage}{ILTER sites
(more than 1200 around the world)},
through the DEIMS-SDR API.
}
\examples{
\dontrun{
listEnvCharacts <- get_ilter_envcharacts(sitesNum = 10)
listEnvCharacts[1:20, ]
}

}
\author{
Alessandro Oggioni, phD (2020) <oggioni.a@irea.cnr.it>
}
