% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_ilter_research_topics.R
\name{get_ilter_research_topics}
\alias{get_ilter_research_topics}
\title{Return a tibble object containing Research Topics collected by all
of the eLTER sites (more than 1200 sites), as a stored into
\href{https://deims.org}{DEIMS-SDR}, belonging to International LTER
network (ILTER).}
\usage{
get_ilter_research_topics(sitesNum = NULL)
}
\arguments{
\item{sitesNum}{A `integer`. It is the number of the sites that are
read to get the information. Use this parameters moreover for provide
example of this function.}
}
\value{
The output of the function is a `tibble` containing the research
topics and their URI (Uniform Resource Identifier) of all ILTER sites.
}
\description{
This function allows to obtain the information about the
Research Topics collected in all sites of ILTER, through the DEIMS-SDR
sites API.
}
\examples{
\dontrun{
listResearchTopics <- get_ilter_research_topics(sitesNum = 20)
listResearchTopics[1:10, ] \%>\%
  dplyr::rows_insert(
  dplyr::tibble(
    researchTopicsLabel = "...",
    researchTopicsUri = "..."
  )
)
}

}
\author{
Alessandro Oggioni, phD (2020) \email{oggioni.a@irea.cnr.it}
}
