% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_location_info.R
\name{get_location_info_internal}
\alias{get_location_info_internal}
\title{Obtain the information about of an eLTER location.}
\usage{
get_location_info_internal(locationid, show_map = FALSE)
}
\arguments{
\item{locationid}{A \code{character}. It is the DEIMS ID of location make from
DEIMS-SDR website. DEIMS ID information
\href{https://deims.org/docs/deimsid.html}{here}.
The DEIMS.iD of activity is the URL for the location page.}

\item{show_map}{A \code{boolean}. If TRUE a Leaflet map with occurrences
is shown. Default FALSE.}
}
\value{
The output of the function is a \code{list} with two elements:
\itemize{
\item \code{map} A Leaflet map with the location, if requested with
\code{show_map}.
\item \code{data} A \code{data.frame} with the information about the location.
}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
This function obtains the information about of an eLTER
location (e.g.
\url{https://deims.org/location/12b38f3f-7e72-425a-80c7-7cad35ce4c7b})
provided in \href{https://deims.org/}{DEIMS-SDR catalogue}.
}
\references{
\insertRef{utilsR}{ReLTER}

\insertRef{dplyrR}{ReLTER}

\insertRef{lubridateR}{ReLTER}

\insertRef{unitsR}{ReLTER}
}
\author{
Alessandro Oggioni, phD (2020) \email{oggioni.a@irea.cnr.it}

Paolo Tagliolato, phD \email{tagliolato.p@irea.cnr.it}
}
\keyword{internal}
