% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_network_parameters.R
\name{get_network_parameters}
\alias{get_network_parameters}
\title{Return a tibble object containing Environmental Parameters collected
by all of the eLTER sites belonging to an eLTER Network (e.g.
\href{https://deims.org/networks/7fef6b73-e5cb-4cd2-b438-ed32eb1504b3}{LTER
Italy network}), as a stored into \href{https://deims.org}{DEIMS-SDR}.}
\usage{
get_network_parameters(networkDEIMSID)
}
\arguments{
\item{networkDEIMSID}{A `character`. The DEIMS.iD of network
from DEIMS-SDR website. DEIMS.iD information 
\href{https://deims.org/docs/deimsid.html}{here} and Complete list of ILTER
networks \href{https://deims.org/search?f[0]=result_type:network}{here}.
The DEIMS.iD of network is the URL for the network page.}
}
\value{
The output of the function is a `tibble` containing the list
of parameters and their URI (Uniform Resource Identifier) collected
by the network's sites.
}
\description{
This function obtains a list of the Environmental Parameters
collected on sites in the eLTER Network (e.g. Italy) through the DEIMS-SDR
sites API.
}
\examples{
\dontrun{
listParams <- get_network_parameters(
  networkDEIMSID =
  "https://deims.org/networks/e0f680c2-22b1-4424-bf54-58aa9b7476a0"
)
listParams[1:10, ] \%>\%
dplyr::rows_insert(
  dplyr::tibble(parameterLabel = "...", parameterUri = "...")
)
}

}
\author{
Alessandro Oggioni, phD (2020) \email{oggioni.a@irea.cnr.it}
}
