% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_network_related_resources.R
\name{get_network_related_resources}
\alias{get_network_related_resources}
\title{Obtain information about the Related Resources (dataset and activity) of a
eLTER Network.}
\usage{
get_network_related_resources(networkDEIMSID)
}
\arguments{
\item{networkDEIMSID}{A \code{character}. It is the DEIMS ID of network make
from DEIMS-SDR website. DEIMS ID information
\href{https://deims.org/docs/deimsid.html}{here} and Complete list of ILTER
networks \href{https://deims.org/search?f\link{0}=result_type:network}{here}.
The DEIMS ID of network is the URL for the network page.}
}
\value{
The output of the function is a \code{tibble} containing the related
resources shared by the network's sites.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
This function obtains the Related Resources information (title
and URL), as a stored in \href{https://deims.org/}{DEIMS-SDR catalogue}, of
all eLTER sites belonging to an eLTER Network (e.g.
\href{https://deims.org/networks/7fef6b73-e5cb-4cd2-b438-ed32eb1504b3}{LTER
Italy network}).
}
\examples{
\dontrun{
listRelatedResources <- get_network_related_resources(
  networkDEIMSID =
  "https://deims.org/network/7fef6b73-e5cb-4cd2-b438-ed32eb1504b3"
)
listRelatedResources
}

}
\references{
\insertRef{jsonliteR}{ReLTER}

\insertRef{dplyrR}{ReLTER}
}
\author{
Alessandro Oggioni, phD (2020) \email{oggioni.a@irea.cnr.it}
}
