% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_network_sites.R
\name{get_network_sites}
\alias{get_network_sites}
\title{Retrieve a list of sites in an eLTER Network.}
\usage{
get_network_sites(networkDEIMSID)
}
\arguments{
\item{networkDEIMSID}{A \code{character}. The DEIMS ID of the network from
DEIMS-SDR website. DEIMS ID information
\href{https://deims.org/docs/deimsid.html}{here} and Complete list of
networks \href{https://deims.org/search?f\link{0}=result_type:network}{here}.
The DEIMS ID of network is the URL for the network page.}
}
\value{
The output of the function is a point vector of \code{sf} class
(package sf) of the network's sites.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
This function return a spatial point vector object including
title, date late updated, URI, and coordinates, stored in
\href{https://deims.org/}{DEIMS-SDR catalogue}, of all the eLTER sites
belonging to an eLTER Network (e.g.
\href{https://deims.org/networks/7fef6b73-e5cb-4cd2-b438-ed32eb1504b3}{LTER-
Italy network}).
}
\examples{
\dontrun{
# The sites of LTER-Italy network
listSites <- get_network_sites(
  networkDEIMSID =
    "https://deims.org/network/7fef6b73-e5cb-4cd2-b438-ed32eb1504b3"
)
listSites

# The sites of LTER Europe network
euSites <- get_network_sites(
  networkDEIMSID =
    "https://deims.org/networks/4742ffca-65ac-4aae-815f-83738500a1fc"
)
euSites
}

}
\author{
Alessandro Oggioni, phD (2020) \email{oggioni.a@irea.cnr.it}
}
