% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_network_sites.R
\name{get_network_sites}
\alias{get_network_sites}
\title{eLTER get_network_sites function}
\usage{
get_network_sites(networkDEIMSID)
}
\arguments{
\item{networkDEIMSID}{A `character`. The DEIMS ID of the network from
DEIMS-SDR website. More information about DEIMS network ID from these pages: 
\href{https://deims.org/docs/deimsid.html}{page}, and
\href{https://deims.org/search?f[0]=result_type:network}{page} the
complete list of ILTER networks.}
}
\value{
The output of the function is `SpatialPointsDataFrame` (package sp)
of the network's sites.
}
\description{
This function retrieves a list of sites in the provided network
including title, date late updated, URI, and coordinates
}
\examples{
\dontrun{
listSites <- get_network_sites(
  networkDEIMSID = "https://deims.org/network/7fef6b73-e5cb-4cd2-b438-ed32eb1504b3"
)
listSites[1:10, ]
}

}
\author{
Alessandro Oggioni, phD (2020) \email{oggioni.a@irea.cnr.it}
}
