% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_site_affiliations.R
\name{get_site_affiliations}
\alias{get_site_affiliations}
\title{eLTER get_site_affiliations function}
\usage{
get_site_affiliations(deimsid)
}
\arguments{
\item{deimsid}{A character. The DEIMS.iD of the site from
DEIMS-SDR website. More information about DEIMS.iD in this pages:
\href{https://deims.org/docs/deimsid.html}{page}.}
}
\value{
The output of the function is a `tibble` with main features of the
site and the affiliations information, such as: networks and projects in
which the site is involved.
}
\description{
This function obtains details about an eLTER site
through the DEIMS-SDR sites API.
}
\author{
Alessandro Oggioni, phD (2020) \email{oggioni.a@irea.cnr.it}
}
\keyword{internal}
