% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_site_boundaries.R
\name{get_site_boundaries}
\alias{get_site_boundaries}
\title{eLTER get_site_boundaries function}
\usage{
get_site_boundaries(deimsid, show_map = FALSE)
}
\arguments{
\item{deimsid}{A `character`. The DEIMS ID of the site from
DEIMS-SDR website. DEIMS ID information 
\href{https://deims.org/docs/deimsid.html}{here}.}

\item{show_map}{a `boolean`. When TRUE the boundary will be plotted on a
Leaflet map. Default FALSE.}
}
\value{
The output of the function is an `sf` object, the boundary of the
site or NA if the boundary is missing from DEIMS-SDR. In addition, as
`html map` with boundaries of the site is plotted.
}
\description{
This function retrieves the boundary of a specified eLTER site
and a view of the site boundaries on a leaflet map is shown.
All the info are taken from the DEIMS-SDR sites API.
If the boundary is missing, a warning message is printed in the R console.
}
\author{
Alessandro Oggioni, phD (2020) \email{oggioni.a@irea.cnr.it}

Micha Silver, phD (2021) \email{silverm@post.bgu.ac.il}
}
\keyword{internal}
