% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_site_contact.R
\name{get_site_contact}
\alias{get_site_contact}
\title{eLTER get_site_contact function}
\usage{
get_site_contact(deimsid)
}
\arguments{
\item{deimsid}{A \code{character}. It is the DEIMS ID of the site from
DEIMS-SDR website. DEIMS ID information
\href{https://deims.org/docs/deimsid.html}{here}.}
}
\value{
The output of the function is a \code{tibble} with main features of the
site and the contact information, such as: site manager, operation
organization, metadata provider, founding agency and site url.
}
\description{
This function obtains the contact information for an
eLTER site through the DEIMS-SDR sites API.
}
\author{
Alessandro Oggioni, phD (2020) \email{oggioni.a@irea.cnr.it}
}
\keyword{internal}
