% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_site_EcoDataCube.R
\name{get_site_EcoDataCube}
\alias{get_site_EcoDataCube}
\title{Acquire various raster layers from
\href{https://ecodatacube.eu/}{EcoDataCube EU}
and crop to an eLTER site boundary.}
\usage{
get_site_EcoDataCube(
  deimsid,
  dataset = "",
  dataset_year = NA,
  dataset_month = NA,
  show_map = TRUE
)
}
\arguments{
\item{deimsid}{\code{string}. The DEIMS ID of the site from
DEIMS-SDR website. DEIMS ID information
\href{https://deims.org/docs/deimsid.html}{here}.}

\item{dataset}{\code{string} The requested dataset. One of:
"CHELSA_precip", "clc_2017", "clc_2020", "crop_map", "DTM_30m",
"MODIS_LST_day", "MODIS_LST_night", "NDVI_bimonthly", "NDVI_yearly",
"NDWI_monthly", "soil_type"
If dataset is NA or an empty string, then
a list of the available EcoDataCube layers is printed.
Default is "".
(See Details for explanation of each dataset)}

\item{dataset_year}{\code{string} indicating which year to choose (for multi-year datasets)
Entered as four digits (i.e. '2020')}

\item{dataset_month}{\code{string} indicating which month (for multi-month datasets).
Entered as two digits. i.e. '02', '03', or '11' etc.}

\item{show_map}{\code{Bool} whether to show plot of downloaded raster.
Default TRUE}
}
\value{
The function returns a SpatRaster object (from the \code{terra} package)
of the requested dataset, cropped to the site boundaries.
If the SpatRaster is categorical, it will contain a colortable (from EcoDataCube).
The user should save the raster to disk, if necessary.
i.e.

\code{writeRaster(ds_site, "site_dataset.tif")}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
Download and return a SpatRaster object containing the requested
dataset from \href{https://ecodatacube.eu/}{EcoDataCube EU},
cropped to an eLTER site boundary, which is obtained from the DEIMS-SDR API.
}
\details{
Supported datasets from the EcoDataCube repository include:
\tabular{llcr}{
dataset       \tab full name                                \tab date required      \tab temporal extent             \tab  res. \cr
------------- \tab ---------------------------------------- \tab ------------- \tab ----- \tab --------------------------- \tab ----- \cr
CHELSA_precip \tab CHELSA Monthly accumulated precipitation \tab yes      \tab 2000-01-01 00:00:00 UTC–2019-06-30 00:00:00 UTC \tab 1000 \cr
clc_2017      \tab Corine Landcover (CLC+) 2017-2019        \tab no       \tab 2017-01-01 00:00:00 UTC–2022-12-31 00:00:00 UTC \tab 10   \cr
clc_2020      \tab Corine Landcover (CLC+) 2020-2022        \tab no       \tab 2017-01-01 00:00:00 UTC–2022-12-31 00:00:00 UTC \tab 10   \cr
crop_map      \tab EUCROPMAP Pan-EU year 2022               \tab no       \tab 2022-01-01 00:00:00 UTC–2022-12-31 00:00:00 UTC \tab 10   \cr
DTM_30m       \tab OpenLandMap Ensemble Digital Terrain Model \tab no  \tab 2006-01-01 00:00:00 UTC–2015-12-31 00:00:00 UTC \tab 30   \cr
MODIS_LST_day \tab MOD11A2 monthly land surface temp. (day)   \tab yes \tab 2000-01-01 00:00:00 UTC–2021-12-31 00:00:00 UTC \tab 1000 \cr
MODIS_LST_night\tab MOD11A2 monthly land surface temp. (night)\tab yes \tab 2000-01-01 00:00:00 UTC–2021-12-31 00:00:00 UTC \tab 1000 \cr
NDVI_bimonthly \tab Cloud-free reconstructed Landsat bimonthly NDVI \tab yes \tab 2000-01-01 00:00:00 UTC–2022-12-31 00:00:00 UTC \tab 30   \cr
NDVI_yearly   \tab Cloud free reconstructed yearly Landsat NDVI     \tab yes \tab 2000-01-01 00:00:00 UTC–2022-12-31 00:00:00 UTC \tab 30    \cr
NDWI_monthly  \tab Cloud free reconstructed bi-monthly NDWI (Gao)   \tab yes \tab 2000-01-01 00:00:00 UTC–2022-12-31 00:00:00 UTC \tab 30    \cr
soil_type     \tab AI4SoilHealth: Soil type dominant class          \tab no  \tab 2000-01-01 00:00:00 UTC–2022-12-31 00:00:00 UTC \tab 30    \cr
}
All datasets are georeferenced to the
EPSG:3035 coordinate reference system.
}
\examples{
# Example of TERENO Harsleben
deimsid = "https://deims.org/c945abe4-3d40-46d1-b5d0-33127c35c6ab"
harsleben_ndvi <- get_site_EcoDataCube(
  deimsid = deimsid,
  dataset = "NDVI_bimonthly",
  dataset_year = "2021",
  dataset+month = "06"
)
harsleben_dtm <- get_site_EcoDataCube(
  deimsid = deimsid,
  dataset = "DTM_30m",
)
 \dontrun{
}
}
\references{
\insertRef{dplyrR}{ReLTER}

\insertRef{terraR}{ReLTER}
}
\author{
Micha Silver, phD (2020) \email{silverm@post.bgu.ac.il}

Alessandro Oggioni, phD (2020) \email{oggioni.a@irea.cnr.it}
}
