% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_site_envcharacts.R
\name{get_site_envcharacts}
\alias{get_site_envcharacts}
\title{eLTER get_site_envcharacts function}
\usage{
get_site_envcharacts(deimsid)
}
\arguments{
\item{deimsid}{A `character`. The DEIMS ID of a site from
DEIMS-SDR website. More information about DEIMS I from:
\href{https://deims.org/docs/deimsid.html}{page}.}
}
\value{
The output of the function is a `tibble` with main features of the
site and the environmental characteristics where available,
such as: air temperature, precipitation, biogeographical region, biome,
ecosystem land use, EUNIS habitat, geoBon biome, geology, hydrology, soils and vegetation.
}
\description{
This function obtains Environmental Characteristics 
of an eLTER site through the DEIMS-SDR sites API.
}
\examples{
tSiteEnvCharacts <- get_site_envcharacts(
  deimsid = "https://deims.org/17210eba-d832-4759-89fa-9ff127cbdf6e"
)
tSiteEnvCharacts

}
\author{
Alessandro Oggioni, phD (2021) \email{oggioni.a@irea.cnr.it}
}
\keyword{internal}
