% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_site_info.R
\name{get_site_info}
\alias{get_site_info}
\title{Obtain details about an eLTER site.}
\usage{
get_site_info(
  deimsid,
  categories = NA,
  show_map = FALSE,
  with_locations = FALSE
)
}
\arguments{
\item{deimsid}{A \code{character}. The DEIMS ID of the site from
DEIMS-SDR website. DEIMS ID information
\href{https://deims.org/docs/deimsid.html}{here}.}

\item{categories}{A \code{categories}. This parameter selects which categories
or categories are retrieved and returned in the result.
Possible value are:
'Affiliations', 'Contacts', 'EnvCharacts', 'General',
'Infrastructure', 'observedProperties', 'RelateRes'.
Multiple values can be indicated.
A site's boundary is always returned.}

\item{show_map}{A \code{boolean}. When TRUE a \code{leaflet} map is plotted as side
effect. Default FALSE.}

\item{with_locations}{A \code{boolean}. When TRUE, and only \code{show_map} is TRUE,
all site related locations are showed in the plotted map.
Default FALSE.}
}
\value{
The output of the function is a \code{sf} with the information about
the site.
If the boundary is missing from DEIMS-SDR a \code{tibble} is returned.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
This function obtains information of a single eLTER site,
as a stored in \href{https://deims.org/}{DEIMS-SDR catalogue},
through the DEIMS-SDR API.
}
\examples{
site <- get_site_info(
  deimsid = "https://deims.org/f30007c4-8a6e-4f11-ab87-569db54638fe",
  categories = c("EnvCharacts", "Affiliations"),
  show_map = TRUE,
  with_locations = FALSE
)
site

}
\references{
\insertRef{dplyrR}{ReLTER}

\insertRef{lubridateR}{ReLTER}

\insertRef{unitsR}{ReLTER}

\insertRef{utilsR}{ReLTER}

\insertRef{sfR}{ReLTER}
}
\author{
Alessandro Oggioni, phD (2020) \email{oggioni.a@irea.cnr.it}

Paolo Tagliolato, phD \email{tagliolato.p@irea.cnr.it}
}
