% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_site_info.R
\name{get_site_info}
\alias{get_site_info}
\title{Obtain details about an eLTER site.}
\usage{
get_site_info(deimsid, category = NA)
}
\arguments{
\item{deimsid}{A character. The DEIMS ID of the site from
DEIMS-SDR website. DEIMS ID information
\href{https://deims.org/docs/deimsid.html}{here}.}

\item{category}{A \code{category}. This parameter selects which category
or categories are retrieved and returned in the result.
Possible value are:
'Affiliations', 'Boundaries', 'Contacts', 'EnvCharacts', 'General',
'Infrastructure', 'observedProperties', 'RelateRes'.
Multiple values can be indicated.}
}
\value{
The output of the function is a \code{tibble} with main features of the
site and the selected information, such as: networks and projects in
which the site is involved.
If category 'Boundaries' is indicated an \code{sf} object is returned
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#stable}{\figure{lifecycle-stable.svg}{options: alt='[Stable]'}}}{\strong{[Stable]}}
This function obtains information of a single eLTER site,
as a stored in \href{https://deims.org/}{DEIMS-SDR catalogue},
through the DEIMS-SDR API.
}
\examples{
site <- get_site_info(
  deimsid = "https://deims.org/f30007c4-8a6e-4f11-ab87-569db54638fe",
  category = "Boundaries"
)
site

siteInfo <- get_site_info(
  deimsid = "https://deims.org/f30007c4-8a6e-4f11-ab87-569db54638fe",
  category = c("EnvCharacts", "Affiliations")
)
siteInfo

}
\references{
\insertRef{dplyrR}{ReLTER}

\insertRef{utilsR}{ReLTER}

\insertRef{sfR}{ReLTER}

\insertRef{leafletR}{ReLTER}
}
\author{
Alessandro Oggioni, phD (2020) \email{oggioni.a@irea.cnr.it}
}
