% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_site_info.R
\name{get_site_info}
\alias{get_site_info}
\title{eLTER get_site_info function}
\usage{
get_site_info(deimsid, category = NA)
}
\arguments{
\item{deimsid}{a character. The DEIMS ID of the site from
DEIMS-SDR website. More information about DEIMS ID in this pages:
\href{https://deims.org/docs/deimsid.html}{page}.}

\item{category}{a `category`. This parameter selects which category
or categories are retrieved and returned in the result.
Possible value are:
'Affiliations', 'Boundaries', 'Contacts', 'EnvCharacts', 'General',
'Infrastructure', 'Parameters', 'RelateRes', 'ResearchTop'.
Multiple values can be indicated.}
}
\value{
The output of the function is a `tibble` with main features of the
site and the selected information, such as: networks and projects in
which the site is involved.
If category 'Boundaries' is indicated an `sf` object is returned
}
\description{
This function obtains details about an eLTER site
through the DEIMS-SDR sites API.
}
\examples{
siteInfo <- get_site_info(
  deimsid = "https://deims.org/f30007c4-8a6e-4f11-ab87-569db54638fe",
  category = c("EnvCharacts", "Affiliations")
)
siteInfo

site <- get_site_info(
  deimsid = "https://deims.org/f30007c4-8a6e-4f11-ab87-569db54638fe",
  category = "Boundaries"
)
site

}
\author{
Alessandro Oggioni, phD (2020) \email{oggioni.a@irea.cnr.it}
}
