% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sites_interactive.R
\name{get_sites_interactive}
\alias{get_sites_interactive}
\title{Select sites by choosing a 3d bounding box through an interactive GUI.}
\usage{
get_sites_interactive()
}
\value{
\code{tibble} with selected sites or
a \code{list} with selected bounding box (with slots \code{bbx} and \code{elevation_range}).
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Open a GUI (a shiny widget) to interact with the user for the
selection of a 3D bounding box. The user can visualise the sites within the selected
bounding box and ask for returning the contained sites or simply the bounding box.
}
\examples{
\dontrun{
sites_tbl_sf <- get_sites_interactive()
}
}
\seealso{
\link[shiny:runGadget]{shiny::runGadget}

\code{\link[leaflet.extras:draw]{leaflet.extras::addDrawToolbar()}}
}
\author{
Paolo Tagliolato
}
