% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sites_within_radius.R
\name{get_sites_within_3d_bounding_box}
\alias{get_sites_within_3d_bounding_box}
\title{select sites within a 3d bounding box}
\usage{
get_sites_within_3d_bounding_box(bbox, elevation_range = NULL, show_map = TRUE)
}
\arguments{
\item{bbox}{bounding box}

\item{elevation_range}{elevation range}

\item{show_map}{show the map}
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
select sites within a 3d bounding box
}
