% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sites_within_radius.R
\name{get_sites_within_radius}
\alias{get_sites_within_radius}
\title{Get all sites within a given distance of a point location}
\usage{
get_sites_within_radius(
  lat = 39.1386,
  lon = -8.33305,
  distance_in_meters = 1000,
  elevation_range = c(NULL, NULL),
  show_map = TRUE
)
}
\arguments{
\item{lat}{latitude of radius center}

\item{lon}{longitude of radius center}

\item{distance_in_meters}{distance from point in meters}

\item{elevation_range}{min and max elevation in meters}

\item{show_map}{\code{logical} print a map with retrieved sites (default is TRUE)}
}
\value{
list of deims_id, distance from given coordinates
}
\description{
retrieve all sites within a given distance (in meters) from
the center point (in coordinates lat, lon), and in
the (optionally) specified elevation range.
}
