% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sos_procedure_info.R
\name{get_sos_procedure_info}
\alias{get_sos_procedure_info}
\title{Obtain the information from a sensor.}
\usage{
get_sos_procedure_info(sosURL, procedure)
}
\arguments{
\item{sosURL}{A `character`. The endpoint of the Sensor Observation Service
(SOS) service.}

\item{procedure}{A `character`. It is a procedure/sensor ID.}
}
\value{
The output of the function is a `character` containing attributes
such as name of the sensor, id and their description.
All the information are collected by requests to Sensor Observation
Service (SOS).
}
\description{
`r lifecycle::badge("experimental")`
This function obtains the information (metadata) shared by
procedure/sensor through Sensor Observation Service (SOS).
}
\examples{
FP <- get_sos_procedure_info(
  sosURL = "http://getit.lteritalia.it/observations/service",
  procedure = "http://www.get-it.it/sensors/getit.lteritalia.it/procedure/noOwnerDeclared/noModelDeclared/noSerialNumberDeclared/1286194C-A5DF-11DF-8ED7-1602DFD72097"
)
FP

eurac_monalisa <- get_sos_procedure_info(
  sosURL = "http://monalisasos.eurac.edu/sos/service",
  procedure = "QuantumSensor_nemef2000"
)
eurac_monalisa

# umweltbundesamt <- get_sos_procedure_info(
#   sosURL = "https://ltercwn.umweltbundesamt.at/cwn-sos2/service",
#   procedure = "KAS_M"
# )
# The value 'D:/deployment/projects/cwnsos2node0140100/Tomcat/sensorML/sos2/sml2/KAS_M.xml' of the parameter 'procedure' is invalid


obsProsAir <- get_sos_procedure_info(
  sosURL = "http://getit.lteritalia.it/observations/service",
  procedure = "http://www.get-it.it/sensors/getit.lteritalia.it/procedure/noOwnerDeclared/noModelDeclared/noSerialNumberDeclared/SI000049-1675AirTemp"
)
obsProsAir

NIVA <- get_sos_procedure_info(
  sosURL = "https://hydro-sos.niwa.co.nz/",
  procedure = "Water_Quality_Site"
)
NIVA

}
\author{
Alessandro Oggioni, phD (2020) \email{oggioni.a@irea.cnr.it}

Paolo Tagliolato, PhD (2021) \email{tagliolato.p@irea.cnr.it}
}
