% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sos_procedurelist.R
\name{get_sos_procedurelist}
\alias{get_sos_procedurelist}
\title{eLTER get_sos_procedurelist function}
\usage{
get_sos_procedurelist(sosHost)
}
\arguments{
\item{sosHost}{a `character`. An SOS endpoint (e.g.
\url{http://getit.lteritalia.it/observations/sos/kvp?}).}
}
\value{
The output of the function is a `list` with the name and URI (Uniform
Resource Identifier) of each procedure.
}
\description{
This function lists the procedures
of a Sensor Observations Service (SOS).
}
\examples{
\dontrun{
get_sos_procedurelist(
  sosHost = "http://getit.lteritalia.it/observations/sos/kvp?"
)
}

}
\author{
Alessandro Oggioni, phD (2020) \email{oggioni.a@irea.cnr.it}
}
