% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sos_procedurelist.R
\name{get_sos_procedurelist}
\alias{get_sos_procedurelist}
\title{List the procedures of a Sensor Observations Service (SOS).}
\usage{
get_sos_procedurelist(sosHost)
}
\arguments{
\item{sosHost}{A `list`. An SOS endpoint (e.g.
\url{http://getit.lteritalia.it/observations/sos/kvp?}).}
}
\value{
The output of the function is a `list` with the name and URI
(Uniform Resource Identifier) of each procedure.
}
\description{
Return a list of Procedure (e.g. Method, algorithm, instrument,
sensor, or system of these which may be used in making an observation)
store into a
\href{http://opengeospatial.github.io/e-learning/sos/text/index.html}{SOS
(Sensor Observations Service OGC)}.
}
\examples{
\dontrun{
get_sos_procedurelist(
  sosHost = "http://getit.lteritalia.it/observations/sos/kvp?"
)
}

}
\author{
Alessandro Oggioni, phD (2020) \email{oggioni.a@irea.cnr.it}
}
