% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_sos_procedurelist.R
\name{get_sos_procedurelist}
\alias{get_sos_procedurelist}
\title{List the procedures of a Sensor Observations Service (SOS).}
\usage{
get_sos_procedurelist(sosURL)
}
\arguments{
\item{sosHost}{A \code{list}. An SOS endpoint (e.g.
\url{http://getit.lteritalia.it/observations/sos/kvp?}).}
}
\value{
The output of the function is a \code{list} with the name and URI
(Uniform Resource Identifier) of each procedure.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Return a list of procedures (e.g. method, algorithm,
instrument, sensor, or system which may be used in making observations)
store into a
\href{http://opengeospatial.github.io/e-learning/sos/text/index.html}{SOS
(Sensor Observations Service OGC)}.
}
\examples{
\dontrun{
get_sos_procedurelist(
  sosURL = "http://getit.lteritalia.it/observations/service"
)
}

}
\references{
\insertRef{xml2R}{ReLTER}

\insertRef{xsltR}{ReLTER}

\insertRef{utilsR}{ReLTER}
}
\author{
Alessandro Oggioni, phD (2020) \email{oggioni.a@irea.cnr.it}

Paolo Tagliolato, PhD (2021) \email{tagliolato.p@irea.cnr.it}
}
