% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_zenodo_data.R
\name{get_zenodo_data}
\alias{get_zenodo_data}
\title{Obtain the data from a dataset deposited in Zenodo record.}
\usage{
get_zenodo_data(doi, rdata_exist = TRUE)
}
\arguments{
\item{doi}{A \code{character}. It is the DOI of the Zenodo record.}

\item{rdata_exist}{A \code{logical}. Is the .RData or .rds file in the record
we are questioning? Default TRUE.}
}
\value{
a file(s) containing in the Zenodo record and tibble contains some
metadata of Zenodo record.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#deprecated}{\figure{lifecycle-deprecated.svg}{options: alt='[Deprecated]'}}}{\strong{[Deprecated]}}
The function download the file(s) deposited in Zenodo record and returns
a tibble with metadata.
}
\examples{
\dontrun{
## Not run:

record <- get_zenodo_data(
  doi = "10.5281/zenodo.7041152", # test dataset
  rdata_exist = TRUE
)
record

## End (Not run)
}

}
\seealso{
\code{\link[zen4R:ZenodoManager]{zen4R::ZenodoManager()}}

\code{\link[tidyr:unnest]{tidyr::unnest()}}

\code{\link[tidyr:nest]{tidyr::nest()}}
}
\author{
Alessandro Oggioni, phD \email{oggioni.a@irea.cnr.it}
}
