% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_zenodo_data.R
\name{get_zenodo_data}
\alias{get_zenodo_data}
\title{Obtain the data from a dataset deposited in Zenodo record.}
\usage{
get_zenodo_data(doi, rdata_exist = TRUE)
}
\arguments{
\item{doi}{A \code{character}. It is the DOI of the Zenodo record.}

\item{rdata_exist}{A \code{logical}. Is the .RData or .rds file in the record
we are questioning? Default TRUE.}
}
\value{
a file(s) containing in the Zenodo record.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
The function returns the dataset, or file(s), deposited in Zenodo record.
}
\examples{
\dontrun{
## Not run:

record <- get_zenodo_data(
  doi = "10.5281/zenodo.7041152", # test dataset
  rdata_exist = TRUE
)

## End (Not run)
}

}
\author{
Alessandro Oggioni, phD \email{oggioni.a@irea.cnr.it}
}
