% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_zenodo_data.R
\name{get_zenodo_data}
\alias{get_zenodo_data}
\title{Obtain the data from a dataset deposited in Zenodo record.}
\usage{
get_zenodo_data(doi, rdata_exist = TRUE)
}
\arguments{
\item{doi}{A `character`. It is the DOI of the Zenodo record.}

\item{rdata_exist}{A `logical`. Is the .RData file in the record we are
questioning? Default TRUE.}
}
\value{

}
\description{
`r lifecycle::badge("experimental")`
The function
}
\examples{
get_zenodo_data(
  doi = "10.5281/zenodo.7041152", # test dataset
  rdata_exist = TRUE
)

}
\author{
Alessandro Oggioni, phD (2020) \email{oggioni.a@irea.cnr.it}
}
