% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getActivity.R
\name{getActivity}
\alias{getActivity}
\title{eLTER_getActivity}
\usage{
getActivity(activityid)
}
\arguments{
\item{activityid}{A character. It is the DEIMS iD of activity make from DEIMS-SDR website. More information about DEIMS iD in this \href{https://deims.org/docs/deimsid.html}{page}.}
}
\value{
The output of the function is a `tibble` with main features of the activity make in a site.
}
\description{
This function allows to obtain the info of activity provided in DEIMS-SDR.
}
\examples{
getDataset(activityid = 'https://deims.org/activity/8689b125-ee46-4d09-9e46-640f9c5c6eab')

}
\author{
Alessandro Oggioni, phD (2020) \email{oggioni.a@irea.cnr.it}
}
