% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getActivity.R
\name{getActivity}
\alias{getActivity}
\title{eLTER getActivity function}
\usage{
getActivity(activityid)
}
\arguments{
\item{activityid}{A character. The DEIMS activity ID from
DEIMS-SDR website. More information about DEIMS iD in this
\href{https://deims.org/docs/deimsid.html}{page}.}
}
\value{
The output of the function is a `tibble` with main features of
the activities in a site, and a `leaflet` map plot.
}
\description{
This function obtains the info of activity (actually
 title) provided in DEIMS-SDR.
}
\examples{
activities <- getActivity(activityid = "https://deims.org/activity/8786fc6d-5d70-495c-b901-42f480182845")
map <- leaflet::leaflet(activities) \%>\% 
 leaflet::addTiles() \%>\% 
 leaflet::addPolygons()
print(map)
activities

}
\author{
Alessandro Oggioni, phD (2020) \email{oggioni.a@irea.cnr.it}
}
