% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getActivity.R
\name{getActivity}
\alias{getActivity}
\title{eLTER getActivity function}
\usage{
getActivity(activityid)
}
\arguments{
\item{activityid}{A character. It is the DEIMS iD of activity make from
DEIMS-SDR website. More information about DEIMS iD in this
\href{https://deims.org/docs/deimsid.html}{page}.}
}
\value{
The output of the function is a `tibble` with main features of
the activity make in a site.
}
\description{
This function allows to obtain the info of activity (actually
 only title) provided in DEIMS-SDR.
}
\examples{
getActivity(activityid = "https://deims.org/activity/8786fc6d-5d70-495c-b901-42f480182845")

}
\author{
Alessandro Oggioni, phD (2020) \email{oggioni.a@irea.cnr.it}
}
