% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getDataset.R
\name{getDataset}
\alias{getDataset}
\title{eLTER getDataset function}
\usage{
getDataset(datasetid)
}
\arguments{
\item{datasetid}{A character. It is the DEIMS iD of dataset make from
DEIMS-SDR website. More information about DEIMS iD in this
\href{https://deims.org/docs/deimsid.html}{page}.}
}
\value{
The output of the function is a `tibble` with main features
of the site and the related resources collected by site.
}
\description{
This function allows to obtain the info of dataset provided in
DEIMS-SDR.
}
\examples{
tDataset <- getDataset(datasetid = "https://deims.org/datasets/75a7f938-7c77-11e3-8832-005056ab003f")
tDataset

}
\author{
Alessandro Oggioni, phD (2020) \email{oggioni.a@irea.cnr.it}
}
