% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getILTEREnvCharacts.R
\name{getILTEREnvCharacts}
\alias{getILTEREnvCharacts}
\title{eLTER getILTEREnvCharacts function}
\usage{
getILTEREnvCharacts(sitesNum = NULL)
}
\arguments{
\item{sitesNum}{A `integer`. It is the number of the sites that are 
read to get the information. Use this parameters oreover for provide
example of this function.}
}
\value{
The output of the function is a `tibble` containing all the
Environmental Characteristics of all ILTER's sites.
}
\description{
This function allows to obtain the information about the
Environmental Characteristics of all sites of ILTER, througth the
DEIMS-SDR sites API.
}
\examples{
\donttest
listEnvCharacts <- getILTEREnvCharacts(sitesNum = 10)
listEnvCharacts[1:20, ]
\donttest

}
\author{
Alessandro Oggioni, phD (2020) <oggioni.a@irea.cnr.it>
}
