% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getILTEREnvCharacts.R
\name{getILTEREnvCharacts}
\alias{getILTEREnvCharacts}
\title{eLTER getILTEREnvCharacts function}
\usage{
getILTEREnvCharacts()
}
\arguments{
\item{null}{function witout parameter}
}
\value{
The output of the function is a `tibble` containing all the
Environmental Characteristics of all ILTER's sites.
}
\description{
This function allows to obtain the information about the
Environmental Characteristics of all sites of ILTER, througth the
DEIMS-SDR sites API.
}
\examples{
listEnvCharacts <- getILTEREnvCharacts()
listEnvCharacts[1:5, ]

}
\author{
Alessandro Oggioni, phD (2020) <oggioni.a@irea.cnr.it>
}
