% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getILTEREnvCharacts.R
\name{getILTEREnvCharacts}
\alias{getILTEREnvCharacts}
\title{eLTER_getILTEREnvCharacts}
\usage{
getILTEREnvCharacts()
}
\value{
The output of the function is a dataframe containing all the Environmental Characteristics of the all ILTER's sites.
}
\description{
This function ...
}
\examples{
getILTEREnvCharacts()

}
\author{
Alessandro Oggioni, phD (2020) <oggioniale@gmail.com>
}
