% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getILTERParameters.R
\name{getILTERParameters}
\alias{getILTERParameters}
\title{eLTER_getILTERParameters}
\usage{
getILTERParameters()
}
\value{
The output of the function is ...
}
\description{
This function ...
}
\examples{
getILTERParameters()

}
\author{
Alessandro Oggioni, phD (2020) <oggioni.a@irea.cnr.it>
}
