% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getILTERResearchTopics.R
\name{getILTERResearchTopics}
\alias{getILTERResearchTopics}
\title{eLTER_getILTERResearchTopics}
\usage{
getILTERResearchTopics()
}
\arguments{
\item{null}{function witout parameter}
}
\value{
The output of the function is a `tibble` containing the research topics and their URI (Uniform Resource Identifier) of all ILTER sites.
}
\description{
This function allows to obtain the information about the Research Topics collected in all sites of ILTER, througth the DEIMS-SDR sites API.
}
\examples{
listResearchTopics <- getILTERResearchTopics()
knitr::kable(listResearchTopics[1:10, ] \%>\% dplyr::rows_insert(tibble(researchTopicsLabel = "...", researchTopicsUri = "...")))

}
\author{
Alessandro Oggioni, phD (2020) \email{oggioni.a@irea.cnr.it}
}
