% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNetworkEnvCharacts.R
\name{getNetworkEnvCharacts}
\alias{getNetworkEnvCharacts}
\title{eLTER_getNetworkEnvCharacts}
\usage{
getNetworkEnvCharacts(networkDEIMSID)
}
\value{
The output of the function is a dataframe containing all the Environmental Characteristics of the network's sites.
}
\description{
This function ...
}
\examples{
getNetworkEnvCharacts(networkDEIMSID = 'https://deims.org/network/7fef6b73-e5cb-4cd2-b438-ed32eb1504b3')

}
\author{
Alessandro Oggioni, phD (2020) <oggioni.a@irea.cnr.it>
}
