% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNetworkParameters.R
\name{getNetworkParameters}
\alias{getNetworkParameters}
\title{eLTER_getNetworkParameters}
\usage{
getNetworkParameters(networkDEIMSID)
}
\value{
The output of the function is a `tibble` containing the list of parameters and their URI (Uniform Resource Identifier) collected by the network's sites.
}
\description{
This function allows to obtain the information about the parameters collected in the eLTER Network (e.g. Italy) througth the DEIMS-SDR sites API.
}
\examples{
listParams <- getNetworkParameters(networkDEIMSID = 'https://deims.org/network/7fef6b73-e5cb-4cd2-b438-ed32eb1504b3')
knitr::kable(listParams[1:10, ] \%>\% dplyr::rows_insert(tibble(parameterLabel = "...", parameterUri = "...")))

}
\author{
Alessandro Oggioni, phD (2020) \email{oggioni.a@irea.cnr.it}
}
