% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNetworkRelatedResources.R
\name{getNetworkRelatedResources}
\alias{getNetworkRelatedResources}
\title{eLTER getNetworkRelatedResources function}
\usage{
getNetworkRelatedResources(networkDEIMSID)
}
\arguments{
\item{networkDEIMSID}{A `character`. It is the DEIMS iD of network make
from DEIMS-SDR website. More information about DEIMS iD in this
\href{https://deims.org/docs/deimsid.html}{page}, and at this
\href{https://deims.org/search?f[0]=result_type:network}{page} the
complete list of ILTER networks.}
}
\value{
The output of the function is a `tibble` containing the related
resources shared by the network's sites.
}
\description{
This function allows to obtain the information about the
related resources (e.g. dataset) shared in the eLTER Network (e.g. Italy)
througth the DEIMS-SDR sites API.
}
\examples{
\dontrun{
listRelatedResources <- getNetworkRelatedResources(
  networkDEIMSID = "https://deims.org/network/7fef6b73-e5cb-4cd2-b438-ed32eb1504b3"
)
listRelatedResources[1:10, ]
}

}
\author{
Alessandro Oggioni, phD (2020) \email{oggioni.a@irea.cnr.it}
}
