% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNetworkRelatedResources.R
\name{getNetworkRelatedResources}
\alias{getNetworkRelatedResources}
\title{eLTER_getNetworkRelatedResources}
\usage{
getNetworkRelatedResources(networkDEIMSID)
}
\value{
The output of the function is ...
}
\description{
This function ...
}
\examples{
getNetworkRelatedResources(networkDEIMSID = 'https://deims.org/network/7fef6b73-e5cb-4cd2-b438-ed32eb1504b3')

}
\author{
Alessandro Oggioni, phD (2020) <oggioni.a@irea.cnr.it>
}
