% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNetworkResearchTopics.R
\name{getNetworkResearchTopics}
\alias{getNetworkResearchTopics}
\title{eLTER getNetworkResearchTopics function}
\usage{
getNetworkResearchTopics(networkDEIMSID)
}
\arguments{
\item{networkDEIMSID}{A `character`. It is the DEIMS iD of network make
from DEIMS-SDR website. More information about DEIMS iD in this
\href{https://deims.org/docs/deimsid.html}{page}, and at this
\href{https://deims.org/search?f[0]=result_type:network}{page} the
complete list of ILTER networks.}
}
\value{
The output of the function is a `tibble` containing the research
topics and their URI (Uniform Resource Identifier) collected by network's
sites.
}
\description{
This function allows to obtain the information about the
Research Topics collected in the eLTER Network (e.g. Italy) througth
the DEIMS-SDR sites API.
}
\examples{
\dontrun{
listResearchTopics <- getNetworkResearchTopics(
  networkDEIMSID = "https://deims.org/network/7fef6b73-e5cb-4cd2-b438-ed32eb1504b3"
)
listResearchTopics[1:10, ] \%>\%
  dplyr::rows_insert(
    dplyr::tibble(
      researchTopicsLabel = "...",
      researchTopicsUri = "..."
    )
  )
}

}
\author{
Alessandro Oggioni, phD (2020) \email{oggioni.a@irea.cnr.it}
}
