% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNetworkSites.R
\name{getNetworkSites}
\alias{getNetworkSites}
\title{eLTER_spatial_points_of_sites}
\usage{
getNetworkSites(networkDEIMSID)
}
\value{
The output of the function is `SpatialPointsDataFrame` (package sp) of the network's sites.
}
\description{
This function provide a SpatialPointsDataFrame (sp) of sites in a network provided by DEIMS-SDR.
}
\examples{
getNetworkSites(networkDEIMSID = 'https://deims.org/network/7fef6b73-e5cb-4cd2-b438-ed32eb1504b3')

}
\author{
Alessandro Oggioni, phD (2020) \email{oggioni.a@irea.cnr.it}
}
