% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getNetworkSites.R
\name{getNetworkSites}
\alias{getNetworkSites}
\title{eLTER getNetworkSites function}
\usage{
getNetworkSites(networkDEIMSID)
}
\arguments{
\item{networkDEIMSID}{A `character`. It is the DEIMS iD of network make from
DEIMS-SDR website. More information about DEIMS iD in this
\href{https://deims.org/docs/deimsid.html}{page}, and at this
\href{https://deims.org/search?f[0]=result_type:network}{page} the
complete list of iLTER networks.}
}
\value{
The output of the function is `SpatialPointsDataFrame` (package sp)
of the network's sites.
}
\description{
This function provide a SpatialPointsDataFrame (sp) of sites in
a network provided by DEIMS-SDR.
}
\examples{
\donttest
listSites <- getNetworkSites(
  networkDEIMSID = "https://deims.org/network/7fef6b73-e5cb-4cd2-b438-ed32eb1504b3"
)
listSites[1:10, ]
\donttest

}
\author{
Alessandro Oggioni, phD (2020) \email{oggioni.a@irea.cnr.it}
}
