% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSiteAffiliations.R
\name{getSiteAffiliations}
\alias{getSiteAffiliations}
\title{eLTER_getSiteAffiliations}
\usage{
getSiteAffiliations(deimsid)
}
\value{
The output of the function is a tibble with main features of the site and the affiliations information, such as: networks and projects in which the site is involved.
}
\description{
This function allows to obtain the information about affiliations of the eLTER site througth the DEIMS-SDR sites API.
}
\examples{
getSiteAffiliations(deimsid = 'https://deims.org/17210eba-d832-4759-89fa-9ff127cbdf6e')

}
\author{
Alessandro Oggioni, phD (2020) <oggioni.a@irea.cnr.it>
}
