% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSiteAffiliations.R
\name{getSiteAffiliations}
\alias{getSiteAffiliations}
\title{eLTER getSiteAffiliations function}
\usage{
getSiteAffiliations(deimsid)
}
\arguments{
\item{deimsid}{A character. It is the DEIMS iD of network make from DEIMS-SDR website. More information about DEIMS iD in this \href{https://deims.org/docs/deimsid.html}{page}.}
}
\value{
The output of the function is a `tibble` with main features of the site and the affiliations information, such as: networks and projects in which the site is involved.
}
\description{
This function allows to obtain the information about affiliations of the eLTER site througth the DEIMS-SDR sites API.
}
\examples{
tSiteAffiliation <- getSiteAffiliations(deimsid = 'https://deims.org/f30007c4-8a6e-4f11-ab87-569db54638fe')
tSiteAffiliation

}
\author{
Alessandro Oggioni, phD (2020) \email{oggioni.a@irea.cnr.it}
}
