% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSiteBoundaries.R
\name{getSiteBoundaries}
\alias{getSiteBoundaries}
\title{eLTER_getSiteBoundaries}
\usage{
getSiteBoundaries(deimsid)
}
\value{
The output of the function is a xxx with boundaries of the site.
}
\description{
This function allows to obtain a gpkg and a shp version (saved in the actual workdir) of the eLTER site boundaries, a view of the site boundaries on a leaflet map is also showed. All the info are taken from the DEIMS-SDR sites API. If the boundaries is missing an worning message will be proposed in the R console.
}
\examples{
getSiteBoundaries(deimsid = 'https://deims.org/f30007c4-8a6e-4f11-ab87-569db54638fe')

}
\author{
Alessandro Oggioni, phD (2020) <oggioni.a@irea.cnr.it>
}
