% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSiteBoundaries.R
\name{getSiteBoundaries}
\alias{getSiteBoundaries}
\title{eLTER getSiteBoundaries function}
\usage{
getSiteBoundaries(deimsid)
}
\arguments{
\item{deimsid}{A `character`. It is the DEIMS iD of network make from
DEIMS-SDR website. More information about DEIMS iD in this
\href{https://deims.org/docs/deimsid.html}{page}.}
}
\value{
The output of the function is a
`html map` with boundaries of the site.
}
\description{
This function allows to obtain html map of the eLTER site
boundaries, a view of the site boundaries on a leaflet map is also
showed. All the info are taken from the DEIMS-SDR sites API. If the
boundaries is missing an warning message will be proposed in the R console.
}
\examples{
tSiteBoundaries <- getSiteBoundaries(
  deimsid = "https://deims.org/f30007c4-8a6e-4f11-ab87-569db54638fe"
)
tSiteBoundaries

eisenwurzen <- getSiteBoundaries(
  deimsid = "https://deims.org/d0a8da18-0881-4ebe-bccf-bc4cb4e25701"
)
eisenwurzen

}
\author{
Alessandro Oggioni, phD (2020) \email{oggioni.a@irea.cnr.it}
}
