% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSiteBoundaries.R
\name{getSiteBoundaries}
\alias{getSiteBoundaries}
\title{eLTER getSiteBoundaries function}
\usage{
getSiteBoundaries(deimsid)
}
\arguments{
\item{deimsid}{A `character`. The DEIMS ID of the site from
DEIMS-SDR website. More information about DEIMS ID in this page:
\href{https://deims.org/docs/deimsid.html}{page}.}
}
\value{
The output of the function is an `sf` object, the boundary of the site
or NA if the boundary is missing from DEIMS-SDR.
In addition, as `html map` with boundaries of the site is plotted.
}
\description{
This function retrieves the boundary of a specified eLTER site
and a view of the site boundaries on a leaflet map is shown.
All the info are taken from the DEIMS-SDR sites API. 
If the boundary is missing, a warning message is printed in the R console.
}
\examples{
tSiteBoundaries <- getSiteBoundaries(
  deimsid = "https://deims.org/f30007c4-8a6e-4f11-ab87-569db54638fe"
)
mapA <- leaflet::leaflet(tSiteBoundaries) \%>\% 
 leaflet::addTiles() \%>\% 
 leaflet::addPolygons()
print(mapA)
tSiteBoundaries

eisenwurzen <- getSiteBoundaries(
  deimsid = "https://deims.org/d0a8da18-0881-4ebe-bccf-bc4cb4e25701"
)
mapB <- leaflet::leaflet(eisenwurzen) \%>\% 
 leaflet::addTiles() \%>\% 
 leaflet::addPolygons()
print(mapB)
eisenwurzen

}
\author{
Alessandro Oggioni, phD (2020) \email{oggioni.a@irea.cnr.it}
}
