% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSiteBoundaries.R
\name{getSiteBoundaries}
\alias{getSiteBoundaries}
\title{eLTER getSiteBoundaries function}
\usage{
getSiteBoundaries(deimsid)
}
\arguments{
\item{deimsid}{A `character`. It is the DEIMS iD of network make from
DEIMS-SDR website. More information about DEIMS iD in this
\href{https://deims.org/docs/deimsid.html}{page}.}
}
\value{
The output of the function is a `gpkg`, a `shapefile` (shp) and a
`html map` with boundaries of the site.
}
\description{
This function allows to obtain a gpkg, a shp and a png
versions (saved in the actual workdir) of the eLTER site boundaries, a view
of the site boundaries on a leaflet map is also showed. All the info are
taken from the DEIMS-SDR sites API. If the boundaries is missing an worning
message will be proposed in the R console.
}
\examples{
tSiteBoundaries <- getSiteBoundaries(
  deimsid = "https://deims.org/f30007c4-8a6e-4f11-ab87-569db54638fe"
)
tSiteBoundaries

}
\author{
Alessandro Oggioni, phD (2020) \email{oggioni.a@irea.cnr.it}
}
