% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSiteEnvCharacts.R
\name{getSiteEnvCharacts}
\alias{getSiteEnvCharacts}
\title{eLTER_getSiteEnvCharacts}
\usage{
getSiteEnvCharacts(deimsid)
}
\value{
The output of the function is a tibble with main features of the site and the environmental characteristics information, such as: air temperature, precipitation, biogeographical region, biome, ecosystem land use, EUNIS habitat, geoBon biome, geology, hydrology, soils and vegetation.
}
\description{
This function allows to obtain the information acout environmental characteristics of the eLTER site througth the DEIMS-SDR sites API.
}
\examples{
getSiteEnvCharacts(deimsid = 'https://deims.org/17210eba-d832-4759-89fa-9ff127cbdf6e')

}
\author{
Alessandro Oggioni, phD (2021) <oggioni.a@irea.cnr.it>
}
