% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSiteGeneral.R
\name{getSiteGeneral}
\alias{getSiteGeneral}
\title{eLTER getSiteGeneral function}
\usage{
getSiteGeneral(deimsid)
}
\arguments{
\item{deimsid}{A `character`. The DEIMS ID of the site from
DEIMS-SDR website. More information about DEIMS ID from:
\href{https://deims.org/docs/deimsid.html}{page}.}
}
\value{
The output of the function is a `tibble` with main features of the
site and the general information, such as: abstract, keywords, purpose,
status, yearEstablished, yearClosed, hierarchy, siteName, short name, site
type, protection level, images.
}
\description{
This function obtains general information 
about an eLTER site througth the DEIMS-SDR sites API.
}
\examples{
tSiteGeneral <- getSiteGeneral(
  deimsid = "https://deims.org/17210eba-d832-4759-89fa-9ff127cbdf6e"
)
tSiteGeneral

}
\author{
Alessandro Oggioni, phD (2020) \email{oggioni.a@irea.cnr.it}
}
