% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSiteInfrastructure.R
\name{getSiteInfrastructure}
\alias{getSiteInfrastructure}
\title{eLTER getSiteInfrastructure function}
\usage{
getSiteInfrastructure(deimsid)
}
\arguments{
\item{deimsid}{A `character`. It is the DEIMS iD of network make from DEIMS-SDR website. More information about DEIMS iD in this \href{https://deims.org/docs/deimsid.html}{page}.}
}
\value{
The output of the function is a `tibble` with main features of the site and the general information, such as: abstract, keywords, porpose, status, yearEstablished, yearClosed, hierarchy, siteName, short name, site type, protection level, images.
}
\description{
This function allows to obtain the information about infrastructure of the eLTER site througth the DEIMS-SDR sites API.
}
\examples{
tSiteInfrastructure <- getSiteInfrastructure(deimsid = 'https://deims.org/17210eba-d832-4759-89fa-9ff127cbdf6e')
tSiteInfrastructure

}
\author{
Alessandro Oggioni, phD (2020) \email{oggioni.a@irea.cnr.it}
}
