% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSiteParameters.R
\name{getSiteParameters}
\alias{getSiteParameters}
\title{eLTER getSiteParameters function}
\usage{
getSiteParameters(deimsid)
}
\arguments{
\item{deimsid}{A `character`. It is the DEIMS iD of network make from
DEIMS-SDR website. More information about DEIMS iD in this
\href{https://deims.org/docs/deimsid.html}{page}.}
}
\value{
The output of the function is a `tibble` with main features of the
site and the parameters collected by site.
}
\description{
This function allows to obtain the list of parameters measured
in the eLTER site througth the DEIMS-SDR sites API.
}
\examples{
tSiteParameters <- getSiteParameters(
  deimsid = "https://deims.org/f30007c4-8a6e-4f11-ab87-569db54638fe"
)
tSiteParameters

}
\author{
Alessandro Oggioni, phD (2020) \email{oggioni.a@irea.cnr.it}
}
