% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSiteParameters.R
\name{getSiteParameters}
\alias{getSiteParameters}
\title{eLTER_getSiteParameters}
\usage{
getSiteParameters(deimsid)
}
\value{
The output of the function is a tibble with main features of the site and the parameters collected by site.
}
\description{
This function allows to obtain the list of parameters measured in the eLTER site througth the DEIMS-SDR sites API.
}
\examples{
getSiteParameters(deimsid = 'https://deims.org/f30007c4-8a6e-4f11-ab87-569db54638fe')

}
\author{
Alessandro Oggioni, phD (2020) <oggioni.a@irea.cnr.it>
}
