% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSiteRelatedResources.R
\name{getSiteRelatedResources}
\alias{getSiteRelatedResources}
\title{eLTER_getSiteRelatedResources}
\usage{
getSiteRelatedResources(deimsid)
}
\arguments{
\item{deimsid}{A `character`. It is the DEIMS iD of network make from DEIMS-SDR website. More information about DEIMS iD in this \href{https://deims.org/docs/deimsid.html}{page}.}
}
\value{
The output of the function is a `tibble` with main features of the site and the related resources collected by site.
}
\description{
This function allows to obtain the list of related resources collected in the eLTER site througth the DEIMS-SDR sites API.
}
\examples{
getSiteRelatedResources(deimsid = 'https://deims.org/ede67a31-079a-4db5-b3a2-83b22054c661')

}
\author{
Alessandro Oggioni, phD (2020) \email{oggioni.a@irea.cnr.it}
}
