% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSiteResearchTopics.R
\name{getSiteResearchTopics}
\alias{getSiteResearchTopics}
\title{eLTER getSiteResearchTopics function}
\usage{
getSiteResearchTopics(deimsid)
}
\arguments{
\item{deimsid}{A `character`. The DEIMS ID of the site from
DEIMS-SDR website. More information about DEIMS ID from:
\href{https://deims.org/docs/deimsid.html}{page}.}
}
\value{
The output of the function is a `tibble` with main features of the
site and a list of the research topics handled in this site.
}
\description{
This function obtains a list of research topics handled 
at an eLTER site througth the DEIMS-SDR sites API.
}
\examples{
tSiteResearchTopics <- getSiteResearchTopics(
  deimsid = "https://deims.org/f30007c4-8a6e-4f11-ab87-569db54638fe"
)
tSiteResearchTopics

}
\author{
Alessandro Oggioni, phD (2020) \email{oggioni.a@irea.cnr.it}
}
