% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSOSProcedureList.R
\name{getSOSProcedureList}
\alias{getSOSProcedureList}
\title{eLTER getSOSProcedureList function}
\usage{
getSOSProcedureList(sosHost)
}
\arguments{
\item{sosHost}{a `character`. An SOS endpoint (e.g. http://getit.lteritalia.it).
In particular is the path before  '/observations/service?'}
}
\value{
The output of the function is a `list` with the name and URI (Uniform
Resource Identifier) of each procedure.
}
\description{
This function lists the procedures 
of a Sensor Observations Service (SOS).
}
\examples{
getSOSProcedureList(sosHost = "http://getit.lteritalia.it")

}
\author{
Alessandro Oggioni, phD (2020) \email{oggioni.a@irea.cnr.it}
}
