% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getSOSProcedureList.R
\name{getSOSProcedureList}
\alias{getSOSProcedureList}
\title{eLTER_getNetworkParameters}
\usage{
getSOSProcedureList(sosHost)
}
\value{
The output of the function is ...
}
\description{
This function ...
}
\examples{
getSOSProcedureList(sosHost = 'https://...')

}
\author{
Alessandro Oggioni, phD (2020) <oggioni.a@irea.cnr.it>
}
