% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mapNetworkPoint.R
\name{mapNetworkPoint}
\alias{mapNetworkPoint}
\title{eLTER mapNetworkPoint function}
\usage{
mapNetworkPoint(networkDEIMSID, countryCode)
}
\arguments{
\item{networkDEIMSID}{A `character`. The DEIMS ID of the network 
from DEIMS-SDR website. More information about DEIMS network ID from:
\href{https://deims.org/docs/deimsid.html}{page}, and
\href{https://deims.org/search?f[0]=result_type:network}{page} 
(the complete list of iLTER networks.)}

\item{countryCode}{A `character` following the ISO 3166-1 alpha-3 codes. This
ISO convention consists of three-letter country codes as defined in ISO 3166-1.
the ISO 3166 standard published by the International Organization for
Standardization (ISO), to represent countries, dependent territories, and
special areas of geographical interest. The map produced by this function
will be limited only to the country indicated in this parameter, if the
network has a extraterritorial sites those will not represented.}
}
\value{
The output of the function is a `tmap` plot containing an image of
geographic distribution of the network of sites present in the chosen country.
}
\description{
This function provides a map (image) of sites in an LTER network
}
\examples{
\dontrun{
# Italian sites
map <- mapNetworkPoint(
  networkDEIMSID = 'https://deims.org/network/7fef6b73-e5cb-4cd2-b438-ed32eb1504b3',
  countryCode = 'ITA'
)
map

# German sites
mapNetworkPoint(
  networkDEIMSID = 'https://deims.org/networks/e904354a-f3a0-40ce-a9b5-61741f66c824',
  countryCode = 'DEU'
)
}

}
\author{
Alessandro Oggioni, phD (2020) \email{oggioni.a@irea.cnr.it}
}
