% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produce_site_parameters_pie.R
\name{produce_site_parameters_pie}
\alias{produce_site_parameters_pie}
\title{Return a pie chart of Environmental Parameters, as a stored in
\href{https://deims.org/}{DEIMS-SDR catalogue}, of a single eLTER site
or all of the eLTER sites belonging to an eLTER Network (e.g.
\href{https://deims.org/networks/7fef6b73-e5cb-4cd2-b438-ed32eb1504b3}{LTER
Italy network}).}
\usage{
produce_site_parameters_pie(deimsid)
}
\arguments{
\item{deimsid}{A `character`. It is the DEIMS ID of site/network from
DEIMS-SDR website. More information about DEIMS ID from:
\href{https://deims.org/docs/deimsid.html}{page}.}
}
\value{
The output of the function is a pie chart.
}
\description{
This function produces a pie chart of the parameters
collected in a site or network grouped into compounds.
}
\examples{
\dontrun{
pie <- produce_site_parameters_pie(
  deimsid = "https://deims.org/f30007c4-8a6e-4f11-ab87-569db54638fe"
)
pie
}

}
\author{
Alessandro Oggioni, phD (2020) \email{oggioni.a@irea.cnr.it}
}
