% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produce_site_parameters_waffle.R
\name{produce_site_parameters_waffle}
\alias{produce_site_parameters_waffle}
\title{Produce a waffle chart of the parameters collected in a site LTER.}
\usage{
produce_site_parameters_waffle(deimsid)
}
\arguments{
\item{deimsid}{A `character`. The DEIMS.iD of site/network from:
DEIMS-SDR website. DEIMS.iD information 
\href{https://deims.org/docs/deimsid.html}{here}.}
}
\value{
The output of the function is a waffle chart and a `tibble`. Each
of the squares represents a parameters measured into the selected eLTER
site. The parameters with the same color belong to the same group (e.g.
biological, atmospheric, etc.).
}
\description{
Return a waffle chart of Environmental Parameters, as a stored in
\href{https://deims.org/}{DEIMS-SDR catalogue}, of a single eLTER site.
}
\examples{
\dontrun{
waffle <- produce_site_parameters_waffle(
  deimsid = "https://deims.org/f30007c4-8a6e-4f11-ab87-569db54638fe"
)
waffle
}

}
\author{
Alessandro Oggioni, phD (2020) \email{oggioni.a@irea.cnr.it}
}
