% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produce_site_parameters_waffle.R
\name{produce_site_parameters_waffle}
\alias{produce_site_parameters_waffle}
\title{Return a waffle chart of Environmental Parameters, as a stored in
\href{https://deims.org/}{DEIMS-SDR catalogue}, of a single eLTER site or
all of the eLTER sites belonging to an eLTER Network (e.g.
\href{https://deims.org/networks/7fef6b73-e5cb-4cd2-b438-ed32eb1504b3}{LTER
Italy network}).}
\usage{
produce_site_parameters_waffle(deimsid)
}
\arguments{
\item{deimsid}{A `character`. The DEIMS.iD of site/network from:
DEIMS-SDR website. DEIMS.iD information 
\href{https://deims.org/docs/deimsid.html}{here}.}
}
\value{
The output of the function is a waffle chart.
}
\description{
This function produces a waffle chart of the Environmental
Parameters collected in a site or network grouped into clusters.
}
\examples{
\dontrun{
waffle <- produce_site_parameters_waffle(
  deimsid = "https://deims.org/f30007c4-8a6e-4f11-ab87-569db54638fe"
)
print(waffle)
}

}
\author{
Alessandro Oggioni, phD (2020) \email{oggioni.a@irea.cnr.it}
}
