% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/produce_zenodo_record.R
\name{produce_zenodo_record}
\alias{produce_zenodo_record}
\title{eLTER produce_zenodo_record function}
\usage{
produce_zenodo_record(
  mytoken,
  myfile,
  dataset_type = "dataset",
  dataset_title,
  dataset_description,
  dataset_authors,
  dataset_license = "CC-BY-SA-4.0",
  dataset_accessRight = "open",
  dataset_version = "1.0",
  dataset_language = "eng",
  dataset_keywords,
  dataset_relatedIdentifier,
  dataset_communities = "lter-italy",
  dataset_grants,
  dataset_publish = FALSE
)
}
\arguments{
\item{mytoken}{a `character`. Scopes assign permissions to your personal
access token. A personal access token works just like a normal OAuth access
token for authentication against the API. This token can be created
\href{https://zenodo.org/account/settings/applications/tokens/new/}{here}.}

\item{myfile}{a `character`. A file for a given Zenodo deposited record.}

\item{dataset_type}{a `character`. It is the type of the record to be
upload. Currently only "dataset" is possible, because the aim of the
function is create a record of dataset in Zenodo repository. Default
"dataset".}

\item{dataset_title}{a `character`. The title of the dataset.}

\item{dataset_description}{a `character`. The description of the dataset.}

\item{dataset_authors}{a `tibble`. It is a list of creator(s) of the
dataset. The approach is to use the firstname, lastname, affiliation,
orcid of the authors of the dataset.}

\item{dataset_license}{a `character`. Only one value is possible. It is the
license with which the dataset is released. The license should be set
with the Zenodo id of the license. Default "CC-BY-SA-4.0".}

\item{dataset_accessRight}{a `character`. Only one value is possible.
Default "open".}

\item{dataset_version}{a `character`. It is a version of the dataset. Only
one value is possible. Default "1.0".}

\item{dataset_language}{a `character`. It is the language of the dataset.
Only one value is possible. Default "eng".}

\item{dataset_keywords}{a `character`. A multiple values are possible. The
keyword to the record of dataset.}

\item{dataset_relatedIdentifier}{a `tibble`. It is the related entities
(e.g. scientific paper, data paper, etc.) with the detaset. The tibble is
composed by 2 variables: relation and identifier. Relation can be one of
among following values: "isCitedBy", "cites", "isSupplementTo",
"isSupplementedBy", "isNewVersionOf", "isPreviousVersionOf",
"isPartOf", "hasPart", "compiles", "isCompiledBy", "isIdenticalTo",
"isAlternateIdentifier". While identifier is any type of identifier
(e.g. DOI).}

\item{dataset_grants}{a `character`. A multiple values are possible. Put a
list of project ids as well as showed by the European Commission via
\href{http://www.openaire.eu/}{OpenAIRE}.}

\item{dataset_publish}{a `boolean`. State whether the Zenodo record is to
be published. The parameter publish can be set to TRUE (to use CAUTIOUSLY,
only if you want to publish your record). Default FALSE.}

\item{dataset_community}{a `character`. It is a name of community as created
in Zenodo. A multiple values are possible. Default "lter-italy".}
}
\value{
A link (URL) of the deposited record in Zenodo. The user must then:
1. visit the page of record, 2. reserve DOI by click on the button
'Reserve DOI', 3. check the information provided, and 4. save and publish
the dataset on Zenodo repository.
}
\description{
This function allows to deposit a dataset to
\href{https://zenodo.org}{Zenodo} repository. The function use the facility
of `zen4r`package. Blondel, Emmanuel, & Barde, Julien. (2021). zen4R: R
Interface to Zenodo REST API (0.5-2). Zenodo.
\url{https://doi.org/10.5281/zenodo.5741143}.
}
\examples{
\dontrun{
# This is only an example!
authors <- tibble::tibble(
  name = c("Luke", "Leia"),
  surname = c("Skywalker", "Organa"),
  affiliation = c("Tatooine", "Alderaan"),
  orcid = c("0000-0002-7997-219X", "0000-0002-7997-219X")
)
keywords <- c("Star Wars", "species", "films", "planets")
relatedIdentifiers <- tibble::tibble(
  relation = c("isSupplementTo", "isPartOf"),
  identifier = c("10.1038/s4150-01-0032", "10.1016/j.2051.06.026")
)
grants <- c("871128", "654359", "871126")
produce_zenodo_record(
  # mytoken = "bWFNNtanuk8ys",
  mytoken = 'bWFNNtanuk6ysFb5jSaIfGjMv8OlM2X3r2vMURIt2kKvfP795tIiuL3frv9e',
  myfile = "data/meteo_data.csv",
  dataset_title = "The title",
  dataset_description = "This is the description of the dataset.",
  dataset_authors = authors,
  dataset_keywords = keywords,
  dataset_relatedIdentifier = relatedIdentifiers,
  dataset_grants = grants,
  dataset_publish = FALSE
)
}

}
\author{
Alessandro Oggioni, phD (2020) \email{oggioni.a@irea.cnr.it}
}
