% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eLTER_reporting.R
\name{reporting_produce_data_object_v1.3}
\alias{reporting_produce_data_object_v1.3}
\title{Compose an eLTER Data Reporting Format object}
\usage{
reporting_produce_data_object_v1.3(
  data = NULL,
  station = NULL,
  method = NULL,
  reference = NULL,
  event = NULL,
  sample = NULL,
  licence = "",
  deimsid = "",
  data_type = "measurement",
  filename = NULL
)
}
\arguments{
\item{data}{A `tibble`. See eLTER data specification format for details}

\item{station}{A `tibble`}

\item{method}{A `tibble`}

\item{reference}{A `tibble`}

\item{event}{A `tibble`}

\item{sample}{A `tibble`}

\item{licence}{A `character`}

\item{data_type}{A `character`}

\item{filename}{optional filename associated with the object, of the form
provided as output by the function `reporting_compose_file_name`}
}
\value{
list with eLTER reporting format slots
}
\description{
Given several tables, creates an eLTER data reporting format
object
}
\note{
This method must be intended as a signpost for future implementation
}
\examples{
\dontrun{
## Not run:

deimsid <- "https://deims.org/8eda49e9-1f4e-4f3e-b58e-e0bb25dc32a6"
time_span <- 2015 # e.g. whole year
# time_span <- "20150302-20180415" # e.g. span between two dates
data_topic <- "VEG" # data provider defined abbreviation of "vegetation"
variable_group <- "SPECCOVER" # data provider defined abbreviation
version <- "V20220907"

filename <- reporting_compose_file_name(
  deimsid = deimsid,
  data_topic = data_topic,
  variable_group = variable_group,
  time_span = time_span,
  version = version
)

data <- dplyr::tribble(
  ~id, ~value,
  1, 7.5,
  2, 4.2
)
station <- dplyr::tribble(
  ~SITE_CODE, ~STATION_CODE, ~STYPE, ~LAT,      ~LON,       ~ALTITUDE,
  deimsid,    "IP2",         "AREA",  45.340805, 7.88887495, 265
)
method <- dplyr::tribble(
  ~VARIABLE, ~METH_DESCR,
  "COVE_F",  "Analysis of ammonium..." 
)

research_object <- reporting_produce_data_object_v1.3(
 filename = filename,
 deimsid = deimsid,
 data = data,
 station = station,
 method = method
)

}
## End (Not run)

}
\seealso{
Peterseil, Geiger et al. (2020)
Field Specification for data reporting. Technical Document.
TechDoc.01. EU Horizon 2020 eLTER PLUS Project, Grant agreement No. 871128
\url{https://zenodo.org/record/6373410}
}
\author{
Paolo Tagliolato, phD \email{tagliolato.p@irea.cnr.it}

Alessandro Oggioni, phD \email{oggioni.a@irea.cnr.it}
}
