% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eLTER_reporting.R
\name{reporting_produce_data_object_v2.0}
\alias{reporting_produce_data_object_v2.0}
\title{Compose an eLTER Data Reporting Format object}
\usage{
reporting_produce_data_object_v2.0(
  station = NULL,
  method = NULL,
  data = NULL,
  reference = NULL,
  event = NULL,
  sample_event = NULL,
  license = NULL,
  deimsid = NULL,
  data_topic,
  variable_group = "",
  time_span,
  version = Sys.Date() \%>\% format("V\%Y\%m\%d")
)
}
\arguments{
\item{station}{A \code{tibble} containing the station information. Station is an
observation entity within a site or platform. Station in this respect is
synonym to plot, observation location, sensor location, etc. and is defined
by a location, elevation and installation height (if relevant) which is
located within a given LTER site or platform.}

\item{method}{A \code{tibble} containing the method information. Method describes
the procedure to generate and manipulate the data.}

\item{data}{A \code{tibble} containing the data. Data are defined as the sum of
observation values being observed at a station/plot either by sensor,
measurement device or human observation.}

\item{reference}{A \code{tibble} containing the reference information. Reference
is the listing and description of additional codes used in the data
provision.}

\item{event}{A \code{tibble} containing the event information. Event is defined
as activity to observe or collect information on the ecosystem
characteristic of interest.}

\item{sample_event}{A \code{tibble} containing the sample event information.
Sample event is the key observational units in environmental sciences, i.e.
ecology, geosciences, biogeochemistry, and hydrobiology, and are essential
to document and further analyse biological communities in laboratories
(e.g., phytoplankton communities in water samples, benthic communities in
sediments, etc.).}

\item{license}{A \code{character}. It is a textual description of the conditions
to use for the data.}

\item{deimsid}{A \code{character} The DEIMS ID of the site from
DEIMS-SDR website. More information about DEIMS ID in this pages:
\href{https://deims.org/docs/deimsid.html}{page}.}

\item{data_topic}{A \code{character}. Max 5-digit code for data topic or
observation programme, e.g. METEO (Meteorology), BIODIV (Biodiversity),
DEPO (deposition), GHG (Green House gas), SW (Soil water), VEG (Vegetation).
The abbreviation is defined by the data provider depending on the data.}

\item{variable_group}{A \code{character}. Optional, list of variables or
variable groups contained in the data. The abbreviation is defined by the
data provider depending on the data.}

\item{time_span}{A \code{numeric} or a \code{character}. Time span covered in the
data. E.g. 2015, 20150302-20180415. The time span is defined by the data
provider.}

\item{version}{version in format "VYYYYMMDD". Data version in the format
“V”YYYYMMDD.
Defaults to current system date.}
}
\value{
list with eLTER reporting format slots
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Given several tables, creates an eLTER data reporting format object
}
\examples{
\dontrun{
## Not run:
deimsid <- "https://deims.org/8eda49e9-1f4e-4f3e-b58e-e0bb25dc32a6"
time_span <- 2015 # e.g. whole year
# time_span <- "20150302-20180415" # e.g. span between two dates
data_topic <- "VEG" # data provider defined abbreviation of "vegetation"
variable_group <- "SPECCOVER" # data provider defined abbreviation
version <- "V20220907"

data <- tibble::tribble(
                                                   ~SITE_CODE,   ~VARIABLE,        ~TIME,   ~VALUE,    ~UNIT,
     "https://deims.org/8eda49e9-1f4e-4f3e-b58e-e0bb25dc32a6",      "TEMP", "2016-03-15",    "5.5",     "°C",
     "https://deims.org/8eda49e9-1f4e-4f3e-b58e-e0bb25dc32a6",      "PREC", "2016-03-03",   "10.2",     "mm",
     "https://deims.org/8eda49e9-1f4e-4f3e-b58e-e0bb25dc32a6",      "TEMP", "2016-02-15",    "2.5",     "°C",
     "https://deims.org/8eda49e9-1f4e-4f3e-b58e-e0bb25dc32a6",      "NH4N",    "2016-03",    "5.5",   "mg/l",
     "https://deims.org/8eda49e9-1f4e-4f3e-b58e-e0bb25dc32a6",      "SO4S",    "2016-03",   "10.2",   "mg/l",
     "https://deims.org/8eda49e9-1f4e-4f3e-b58e-e0bb25dc32a6",        "CA",    "2016-03",    "2.5",   "mg/l"
)
station <- dplyr::tribble(
  ~SITE_CODE, ~STATION_CODE, ~STYPE, ~LAT,      ~LON,       ~ALTITUDE,
  deimsid,    "IP2",         "AREA",  45.340805, 7.88887495, 265
)
method <- dplyr::tribble(
  ~VARIABLE, ~METH_DESCR,
  "COVE_F",  "Analysis of ammonium..." 
)

research_object <- reporting_produce_data_object_v2.0(
  station = station,
  method = method,
  data = data,
  deimsid = deimsid,
  data_topic = data_topic,
  variable_group = variable_group,
  time_span = time_span,
  version = version
)

}
## End (Not run)

}
\seealso{
Peterseil, Geiger et al. (2020)
Field Specification for data reporting. Technical Document.
TechDoc.01. EU Horizon 2020 eLTER PLUS Project, Grant agreement No. 871128
\url{https://zenodo.org/record/6373410}
}
\author{
Alessandro Oggioni, phD \email{oggioni.a@irea.cnr.it}
}
