% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eLTER_reporting.R
\name{reporting_save_archive}
\alias{reporting_save_archive}
\title{Creates an archive with files following the eLTER reportingFormat}
\usage{
reporting_save_archive(x, filepath = tempdir(), saveRDS = FALSE)
}
\arguments{
\item{x}{A \code{list} like the one created by function
\code{reporting_produce_data_object_v2.0}}

\item{filepath}{A \code{character} file path. Defaults to temporary directory}

\item{saveRDS}{A \code{logical}. Save also object in RDS format.
Defaults to FALSE}
}
\value{
named A \code{list} containing paths to saved files filepaths.
Slots are named "zip" and possibly "RDS". A path to the zip file, or rds
file, can be defined by the parameter \code{filepath}.
}
\description{
\ifelse{html}{\href{https://lifecycle.r-lib.org/articles/stages.html#experimental}{\figure{lifecycle-experimental.svg}{options: alt='[Experimental]'}}}{\strong{[Experimental]}}
Creates a zip archive "filename".zip
}
\note{
This method must be intended as a signpost for future implementation
}
\examples{
\dontrun{
## Not run:

archive <- reporting_save_archive(
  x = research_object,
  # obtained from the function `reporting_produce_data_object_v2.0()`
  filepath = ".",
  saveRDS = TRUE
)

## End (Not run)
}

}
\seealso{
Peterseil, Geiger et al. (2020)
Field Specification for data reporting. Technical Document.
TechDoc.01. EU Horizon 2020 eLTER PLUS Project, Grant agreement No. 871128
\url{https://zenodo.org/record/6373410}
}
\author{
Paolo Tagliolato, phD \email{tagliolato.p@irea.cnr.it}

Alessandro Oggioni, phD \email{oggioni.a@irea.cnr.it}
}
