% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_id_worms_refine.R
\name{taxon_id_worms_refine}
\alias{taxon_id_worms_refine}
\title{eLTER taxon_id_worms_refine}
\usage{
taxon_id_worms_refine(
  input,
  taxaColumn = 1,
  interaction = FALSE,
  choiceNumber = NULL
)
}
\arguments{
\item{input}{a `tibble`. The output of taxon_ID_Worms.}

\item{taxaColumn}{a `numeric`. The cardinal number of the column where
species list is. Default is `1`.}

\item{interaction}{a `boolean`. When `TRUE` the user can interact with
the choice of one of the proposed matching. Default is `FALSE`.}

\item{choiceNumber}{a `numeric`. The number of selected choice. Default
is `NULL`. When `NULL` no choice is performed.}
}
\value{
In interactive mode or when the `choiceNumber` is `NULL` the
output of the function is a `tibble` with the refinement of the
input, else the function simply returns the unchanged input.
}
\description{
This function provides a refining of the output of
function taxon_id_worms. It prints a list of possible matches
and possibly enables the user to interactively make his choice.
}
\author{
Alessandro Oggioni, phD (2021) \email{oggioni.a@irea.cnr.it}
}
\keyword{internal}
