% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/taxon_id_worms_refine.R
\name{taxon_id_worms_refine}
\alias{taxon_id_worms_refine}
\title{eLTER taxon_id_worms_refine}
\usage{
taxon_id_worms_refine(input, taxaColumn = 1)
}
\arguments{
\item{input}{a `tibble`. The output of taxon_ID_Worms.}

\item{taxaColumn}{a `numeric`. The cardinal number of the column where
species list is. Default is `1`.}
}
\value{

}
\description{
This function provides a refining of the output of
function taxon_ID_WORMS.
}
\examples{
\dontrun{
phytoplankton <- tibble::tibble(
   ID = c(1, 2, 3, 4, 5, 6, 7),
   species = c("Asterionella formosa", "Chrysococcus sp.", "Cryptomonas rostrata", "Dinobryon divergens", "Mallomonas akrokomos", "Melosira varians", "Cryptomonas rostrata")
)
table <- ReLTER::taxon_ID_Worms(input = phytoplankton, taxaColumn = 2, verbose = TRUE, refine = FALSE)
table
taxon_id_worms_refine(input = table, taxaColumn = 2)
}
}
\author{
Alessandro Oggioni, phD (2021) \email{oggioni.a@irea.cnr.it}
}
\keyword{internal}
